% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actions.mouse.ref.db}
\alias{actions.mouse.ref.db}
\title{Actions database for mouse}
\format{
An object of class \code{data.frame} with 4397576 rows and 9 columns.
}
\source{
STRING <https://string-db.org>
}
\usage{
actions.mouse.ref.db
}
\description{
A dataset collects detailed action informations about gene-gene pairs. It is extracted from
STRING database, and has 496206 in total.
}
\keyword{datasets}
