% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-in-fullview.R
\name{AnalyzeClustersInteracts}
\alias{AnalyzeClustersInteracts}
\title{Analyze interaction pairs based on clusters}
\usage{
AnalyzeClustersInteracts(
  fgenes.remapped.all,
  pairs.ref,
  anno.location.ref,
  anno.type.ref,
  user.type.database = NULL,
  restricted.some.genes = NULL,
  restricted.gene.pairs = NULL,
  sub.sel.X.clusters = NULL,
  sub.sel.Y.clusters = NULL,
  sub.sel.exprs.changes = c("Xup.Yup", "Xup.Ydown", "Xdown.Yup", "Xdown.Ydown"),
  sub.sel.X.Location = character(),
  sub.sel.X.Location.score.limit = c("the most confident"),
  sub.sel.Y.Location = character(),
  sub.sel.Y.Location.score.limit = c("the most confident"),
  sub.sel.X.Type = character(),
  sub.sel.Y.Type = character(),
  sub.sel.user.type.colname = NULL,
  sub.sel.X.user.type = NULL,
  sub.sel.Y.user.type = NULL,
  calculation.formula = EvaluateByFunc,
  ind.colname.end.dual = 4
)
}
\arguments{
\item{fgenes.remapped.all}{Data.frame. Informations about feature genes(markers), prefer using return value of 
\code{\link{DataPrep.RemapClustersMarkers}} and similar functions start with \code{DataPrep.} here.}

\item{pairs.ref}{Data.frame. Reference interaction database provided in this package, the list is given as follows: 
\itemize{
 \item For human: \code{pairs.human.experiments.db}, \code{pairs.human.knowledge.db}, \code{pairs.human.prediction.db}
 \item For mouse: \code{pairs.mouse.experiments.db}, \code{pairs.mouse.knowledge.db}, \code{pairs.mouse.prediction.db}
}}

\item{anno.location.ref}{Data.frame. Annotation of subcellular locations for genes provided in this package, the list is given as follows:
\itemize{
  \item For human: \code{anno.location.human.ref.db}
  \item For mouse: \code{anno.location.mouse.ref.db}
}}

\item{anno.type.ref}{Data.frame. Annotation of types(molecular functions) for genes provided in this package, the list is given as follows:
\itemize{
  \item For human: \code{anno.type.human.ref.db}
  \item For mouse: \code{anno.type.mouse.ref.db}
}}

\item{user.type.database}{Data.frame. It specifies one special data.frame that contains user-defined informations.
For generating such database, see \code{Tool.AddUserRestrictDB} for help.}

\item{restricted.some.genes}{Character. Analysis will be restricted in interaction pairs that contain
at least one genes given in this parameter.}

\item{restricted.gene.pairs}{Character or data.frame. Analysis will be restricted in given gene pairs(i.e. interaction pairs). 
The given format is explained in the below, see details for help.}

\item{sub.sel.X.clusters}{[TODO] specific retrict to some clusters in analysis}

\item{sub.sel.Y.clusters}{[TODO]}

\item{sub.sel.exprs.changes}{Character. Use subset of \code{c("Xup.Yup", "Xup.Ydown", "Xdown.Yup", "Xdown.Ydown")}.}

\item{sub.sel.X.Location}{Character. Its value depends on the database used, see details for help.}

\item{sub.sel.X.Location.score.limit}{Character or Integer. The one in \code{character()} will be treated as predefined strategy, while
the one in \code{integer()} will be treated as score limit range. See details for help.}

\item{sub.sel.Y.Location}{Like \code{sub.sel.X.Location}.}

\item{sub.sel.Y.Location.score.limit}{Like \code{sub.sel.X.Location.score.limit}.}

\item{sub.sel.X.Type}{Character. Its value depends on the database used, see details for help.}

\item{sub.sel.Y.Type}{Like \code{sub.sel.X.Type}.}

\item{sub.sel.user.type.colname}{Character. It gives the column name that will be used for user-defined purposes.}

\item{sub.sel.X.user.type}{Its mode depends on what datatype user has defined. The function will accept anything given in this parameter
without any additional check.}

\item{sub.sel.Y.user.type}{Like \code{sub.sel.X.user.type}.}

\item{calculation.formula}{[TODO]}

\item{ind.colname.end.dual}{Integer. Use default value provided only when the pairs.ref database is modified by users.}
}
\value{
A list.
\itemize{
  \item list.clusters: names of all involved clusters.
  \item data.allpairs: a list of all interaction pairs. e.g. cluster: Astrocyte -> cluster: Microglia, if kClustersSplit is "%",
        then, the interaction pair will be named "Astrocyte%Microglia", and use $data.allpairs[["Astrocyte%Microglia"]] to get details.
  \item anno.allpairs: a list of lists. The sublists are 
        \itemize{
          \item location.A: it records locations of A in gene pairs formatted as A-B.
          \item location.B: it records locations of B in gene pairs formatted as A-B.
          \item type.A: it records molecular functions of A in gene pairs formatted as A-B.
          \item type.B: it records molecular functions of B in gene pairs formatted as A-B.
        }
  \item name.allpairs: the names of $data.allpairs
  \item cnt.allpairs: count of each interaction pair in $data.allpairs
  \item strength.allpairs: power of each interaction pair in $data.allpairs
}
}
\description{
This function analyzes count and power of interaction pairs among all clusters.
}
\details{
This function gives as much as possible freedom for users to choose to analyze a subset of interaction pairs.
\itemize{
  \item \code{restricted.some.genes}: (1) Use GO terms or IDs to get some genes.
        (2) Directly specify some genes(use remap.ref.<taxonomy>.dblist to get authorized genenames).
  \item \code{restricted.gene.pairs}: If it is given, the parameter \code{restricted.some.genes} will be ignored. For this parameter, the allowed
        data formats are:

        (1) Character vector with genes in every pair listed in consecutive order, e.g. if \code{c("geneA", "geneB", "geneC", "geneD")} is given, 
        it will be interpreted to be 2 gene pairs: geneA - geneB, geneC - geneD. In this format, the length of the given vector is implied to be 
        even number.

        (2) Data.frame with 2 columns, which records the gene pairs in every row.

  \item \code{sub.sel.exprs.changes}: For interaction pair X-Y, take "Xup.Ydown" for example, and it means the subset that
        gene X is up-regulated and gene Y is down-regulated is selected.
  \item \code{sub.sel.X.Location} & \code{sub.sel.Y.Location}: Location of genes, i.e. Extracellular Region, Plasma Membrane, Cytosol, etc. 
        To fetch available values from current used database, \bold{e.g.}

        In human, use database \code{anno.location.human.ref.db}, and use the code \code{levels(factor(anno.location.human.ref.db$GO.Term.target))} to 
        see all available values.        
  \item \code{sub.sel.X.Location.score.limit} & \code{sub.sel.Y.Location.score.limit}: Further restrition applied on location of genes, which can be given 
        in \code{character()} or \code{integer()}.

        In \code{character()}, this gives the predefined strategies, which support only 1 strategy in current setting. The "the most confident" strategy, 
        which selects all locations of the highest score for one specific gene, and it means multiply locations can be selected together if their scores are 
        the same.

        In \code{integer()}, this gives the range of score. The full range of score is 1~5, and other integers out of this range will not be recognized, 
        \bold{e.g.} \code{sub.sel.X.Location.score.limit = c(4,5)} is fine, and \code{sub.sel.X.Location.score.limit = c(4,5,6)} has the same meaning as the former 
        one when 6 is out of range and will not be used.
        \bold{TO NOTE}: BOTH strategies will have some effect on final runout data, e.g if use \code{.score.limit = c(5)}, and get gene A with Plasma Membrane in score 4, 
        then gene A will not be seen as it can located in Plasma Membrane.  
  \item \code{sub.sel.X.Type} & \code{sub.sel.Y.Type}: More specific type definition of genes, i.e. Cytokine, G-protein Coupled Receptor, etc.

        In mouse, use database \code{anno.type.mouse.ref.db}, and use the code \code{levels(factor(anno.type.mouse.ref.db$Keyword.Name))}.
  \item \code{sub.sel.X.user.type} & \code{sub.sel.Y.user.type} + \code{sub.sel.user.type.colname}: Those give a more flexible way for subset selection.
        It uses the database given in another parameter \code{user.type.database}. \code{sub.sel.user.type.colname} selects which column to be used. 
        \code{sub.sel.X.user.type} & \code{sub.sel.Y.user.type} select the specific items in that column, and if any of them were set \code{NULL}, the  
        corresponding part will not add restriction on this user type.
  \item \code{ind.colname.end.dual}: It is the index when column of pairs.ref begin to provide information about the pair itself
         but not for the 2 genes in that pair. In almost all circumstances, this parameter is required not to be modified.
}
}
\examples{
\dontrun{
# restrict to some subset
tmp.gene.pairs <- data.frame(geneA = c("Fas", "Tgfbr2", "Rac1"),
                             geneB = c("Casp8", "Tgfbr1", "Ctnnb1"))
AnalyzeClustersInteracts(fgenes.remapped.all,
  pairs.ref = pairs.mouse.experiments.db,
  restricted.gene.pairs = tmp.gene.pairs,
  sub.sel.X.Location = c("Plasma Membrane"),
  sub.sel.Y.Location = c("Plasma Membrane", "Cytosol")
)
}



}
