% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-part1.R
\name{AnalyzeInterInAction}
\alias{AnalyzeInterInAction}
\title{Analysis Composition of Actions}
\usage{
AnalyzeInterInAction(
  object,
  limits.exprs.change = kexprs.change,
  limits.action.mode = kpred.action.mode,
  limits.action.effect = kpred.action.effect
)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{limits.exprs.change}{Options are 'Xup.Yup', 'Xup.Ydn', 'Xdn.Yup', 'Xdn.Ydn'.
It collects the gene pairs matching to given expression change.}

\item{limits.action.mode}{Options are listed in \code{kpred.action.mode}. It specifies 
the range of used action mode.}

\item{limits.action.effect}{Options are listed in \code{kpred.action.effect}. It specifies 
the range of used action effect.}
}
\value{
A \code{InterCell} object.
}
\description{
This function gets to group gene pairs by their expression change, action mode and 
action effect.
}
