% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullView.R
\name{AnalyzeInterInFullView}
\alias{AnalyzeInterInFullView}
\title{Analyze interaction network in full view}
\usage{
AnalyzeInterInFullView(
  object,
  sel.clusters.X = NULL,
  sel.clusters.Y = NULL,
  sel.exprs.change = c("Xup.Yup", "Xup.Ydn", "Xdn.Yup", "Xdn.Ydn"),
  sel.some.genes.X = NULL,
  sel.some.genes.Y = NULL,
  sel.genes.option = "intersect",
  sel.gene.pairs = NULL,
  force.process = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{[TODO] InterCell Object}

\item{sel.clusters.X}{[TODO]}

\item{sel.clusters.Y}{[TODO]}

\item{sel.exprs.change}{[TODO]}

\item{sel.some.genes.X}{[TODO]}

\item{sel.some.genes.Y}{[TODO]}

\item{sel.genes.option}{[TODO]}

\item{sel.gene.pairs}{[TODO] data.frame}

\item{force.process}{[TODO]}
}
\value{
A list.
}
\description{
This function analyzes count and power of interaction pairs among all given clusters.
}
\details{

}
\examples{


}
