% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-part3.R
\name{AnalyzeInterSpecificity}
\alias{AnalyzeInterSpecificity}
\title{Find special genes in one pair of interacting clusters}
\usage{
AnalyzeInterSpecificity(
  object,
  to.cmp.cluster.groups = character(),
  extended.search = FALSE,
  uq.cnt.range = c(1:100)
)
}
\arguments{
\item{object}{[TODO]}

\item{to.cmp.cluster.groups}{Character. It defines the cluster groups to be compared.}

\item{extended.search}{[TODO]}

\item{uq.cnt.range}{Integer. It defines the allowed shared-existence count of one gene pairs, which is one specific gene pairs getting 
to exist in several cluster groups.}
}
\value{
List of length 2. With one named "res" saving all result and the other 
named "for.plot.use" saving all needed data for plotting by function \code{GetResult.SummarySpecialGenes}.
}
\description{
This function is used to find special genes in one pair of interacting clusters. Genes are special
if it passes some limitations when comparing to other pairs of interacting clusters.
}
