% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno.location.human.ref.db}
\alias{anno.location.human.ref.db}
\title{Annotation database for subcellular location of human genes}
\format{
An object of class \code{data.frame} with 79628 rows and 10 columns.
}
\source{
<https://compartments.jensenlab.org>, and download the knowledge channel of human databases.
}
\usage{
anno.location.human.ref.db
}
\description{
A dataset contains 79628 rows of subcellular locations, which covers 17646 human genes.
}
\keyword{datasets}
