% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno.location.mouse.ref.db}
\alias{anno.location.mouse.ref.db}
\title{Annotation database for subcellular location of mouse genes}
\format{
An object of class \code{data.frame} with 66569 rows and 10 columns.
}
\source{
<https://compartments.jensenlab.org>, and download the knowledge channel of mouse databases.
}
\usage{
anno.location.mouse.ref.db
}
\description{
A dataset contains 66569 rows of subcellular locations, which covers 15566 mouse genes.
}
\keyword{datasets}
