% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno.type.human.ref.db}
\alias{anno.type.human.ref.db}
\title{Annotation database for molecular function of human genes}
\format{
An object of class \code{data.frame} with 30756 rows and 6 columns.
}
\source{
<https://www.uniprot.org>, and download the Keywords-[molecular function] of genes.
}
\usage{
anno.type.human.ref.db
}
\description{
A dataset contains 30756 rows of molecular functions, which covers 20195 human genes.
}
\keyword{datasets}
