% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno.type.mouse.ref.db}
\alias{anno.type.mouse.ref.db}
\title{Annotation database for molecular function of mouse genes}
\format{
An object of class \code{data.frame} with 25831 rows and 6 columns.
}
\source{
<https://www.uniprot.org>, and download the Keywords-[molecular function] of genes.
}
\usage{
anno.type.mouse.ref.db
}
\description{
A dataset contains 25831 rows of molecular functions, which covers 16655 mouse genes.
}
\keyword{datasets}
