% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkg-parameters.R
\docType{package}
\name{CellTalkDB-package}
\alias{CellTalkDB}
\alias{CellTalkDB-package}
\title{CellTalkDB: Tools for Analyzing Interactions among Cells}
\description{
This tool provides databases covering human and mouse, and provides bunches of functions intended to serve for data analysis of scRNA-seq. Its databases collect a variety of data source, including STRING, NCBI-gene, GO, HGNC, Mesh, etc. Its function on analysis ranges from fullview of cell-cell-interaction to specific interactions between specific clusters.
}
\details{
This package provides databases and tools for scRNA-seq data analysis.
}
\author{
\strong{Maintainer}: Jin Ziyang Developer \email{454785187@qq.com}

Authors:
\itemize{
  \item Shi Ligen Developer \email{slg0904@163.com} [reviewer]
  \item Zhang Xiaotao Developer \email{notknown@tmp.com} [data contributor]
}

}
\keyword{internal}
