% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckSpeciesValidity}
\alias{CheckSpeciesValidity}
\title{Check Species}
\usage{
CheckSpeciesValidity(species)
}
\arguments{
\item{species}{Allowed species are 'human' and 'mouse'.}
}
\description{
This function helps to check species.
}
