% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{setClusterSplit}
\alias{setClusterSplit}
\alias{setClusterSplit,InterCell-method}
\alias{getClusterSplit}
\alias{getClusterSplit,InterCell-method}
\title{Using Cluster Group Split}
\usage{
setClusterSplit(object, ...)

\S4method{setClusterSplit}{InterCell}(object, new.cluster.split)

getClusterSplit(object)

\S4method{getClusterSplit}{InterCell}(object)
}
\arguments{
\item{new.cluster.split}{A new cluster group split, which is either one character or string.}
}
\description{
The \code{setClusterSplit} function is to \bold{set} the cluster group split, 
which is either one charater or string, and used to split 2 cell clusters in one interaction.

The \code{getClusterSplit} function is to \bold{set} the cluster group split, 
which is either one charater or string, and used to split 2 cell clusters in one interaction.
}
\examples{
\dontrun{
  setClusterSplit(object, some.new.cluster.split)
}

\dontrun{
  getClusterSplit(object)
}
}
