% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed-view-in-targets.R
\name{CollectHierachyOnepairClusters}
\alias{CollectHierachyOnepairClusters}
\title{TODO}
\usage{
CollectHierachyOnepairClusters(
  VEinfos,
  limits.exprs.change = c("Cup.Dup", "Cup.Ddn", "Cdn.Dup", "Cdn.Ddn"),
  limits.action.effects = c("A-->B", "A<--B", "A--|B", "A|--B", "A--oB", "Ao--B",
    "A---B")
)
}
\arguments{
\item{VEinfos}{List. It contains informations about vertices and edges, and is exactly return value of
\code{GenerateVEinfos()} or \code{TrimVEinfos()}.}

\item{limits.exprs.change}{[TODO]}

\item{limits.action.effects}{[TODO]}
}
\value{
TODO
}
\description{
TODO
}
\details{
This function is to collect hierachical information inside the raw data.
}
