% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{CreateDBPackObject}
\alias{CreateDBPackObject}
\title{Create InterCellDBPack Object}
\usage{
CreateDBPackObject(species)
}
\arguments{
\item{species}{\pkg{InterCellDB} supports species in either 'human' or 'mouse'.
Other species are not supported to be used for analysis.}
}
\value{
A InterCellDBPack object.
}
\description{
This function is used to create \code{InterCellDBPack-class}.
}
