% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{DataPrep.GenStdGenePairs.from.VEinfos}
\alias{DataPrep.GenStdGenePairs.from.VEinfos}
\title{Generate Gene Pairs in Standard Format}
\usage{
DataPrep.GenStdGenePairs.from.VEinfos(VEinfos, musthave.colnames)
}
\arguments{
\item{VEinfos}{standard storage for one interaction.}

\item{musthave.colnames}{column names that must be included.}
}
\description{
This function generates gene pairs in standard format(in data frame), 
and gets these pairs easier to be compared with others.
}
\details{
The standard format in this package is that gene pairs are maintained in data.frame, and the 2 genes 
participated in each gene pair are recorded in columns named "inter.GeneName.A" and "inter.GeneName.B".
}
