% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preparation.R
\name{DataPrep.ReplaceClusterName}
\alias{DataPrep.ReplaceClusterName}
\title{Replace cluster names for feature genes}
\usage{
DataPrep.ReplaceClusterName(
  markers.all,
  cluster.names.current,
  cluster.names.replace,
  colnames.cluster = "cluster"
)
}
\arguments{
\item{markers.all}{Data.frame. Feature genes that are from return value of \code{Seurat::FindAllMarkers()}.}

\item{cluster.names.current}{Vector. It specifies the names of clusters used currently.}

\item{cluster.names.replace}{Vector. It gives the replaced names of clusters.}

\item{colnames.cluster}{Character. The name of the column that defining the cluster.}
}
\description{
It replaces the old cluster names of feature genes that are in integer generated automatically by \pkg{Seurat} algorithm.
}
\details{
The replace process can be applied on part of clusters, which means the length of parameters 
\code{cluster.names.current} & \code{cluster.names.replace} can be arbitary one with smaller or equal to the count of all clusters.
The replace process is taken one-by-one way, as a result, the length of \code{cluster.names.current} and \code{cluster.names.replace} 
must be the same, or the replacing will not be adequately applied.
}
\examples{
\dontrun{
 DataPrep.ReplaceClusterName(markers.all,
   cluster.names.current = c(0, 1, 2),
   cluster.names.replace = c("Astrocytes", "Microglia", "Endothelial cells"))
}



}
