% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-common.R
\name{DoPartUnique}
\alias{DoPartUnique}
\title{Speed-up doing unique for data.frame}
\usage{
DoPartUnique(xxpairs, cols.select = c(1:2))
}
\arguments{
\item{xxpairs}{Data.frame. Any data.frame object.}

\item{cols.select}{Integer. Specify some columns in integer vector, e.g. c(1:2).
Its length \bold{must be >= 2}, or the function will not work properly.}
}
\description{
This function uses the properties of \code{data.frame} and \code{rownames}, and 
makes partial unique process fast and effective.
}
\details{
When encountering multi-columns tables, e.g. 6 columns or more, \code{unique()} will 
be really slow if it is applied on the whole table. However, in most circumstances, 
there is no need to apply \code{unique()} on all columns, i.e. only do unique on some 
columns, which is exactly the thing this function does.
}
