% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-prep.R
\name{ExtractTargetOnepairClusters}
\alias{ExtractTargetOnepairClusters}
\title{Extract one pair of interaction}
\usage{
ExtractTargetOnepairClusters(
  interact.pairs.acted,
  cluster.name.C,
  cluster.name.D
)
}
\arguments{
\item{interact.pairs.acted}{A list. The return value of \code{\link{AnalyzeClustersInteracts}}.}

\item{cluster.name.C}{Character. Name of one cluster.}

\item{cluster.name.D}{Character. Name of one cluster.}
}
\value{
A list.
\itemize{
  \item clusters.name: names of clusters.
  \item bt.pairs: a list of all interaction pairs between specified 2 clusters (\code{given in clusters.name}).
  \item anno.infos: a list of lists. The sublists are 
        \itemize{
          \item location.A: it records locations of A in gene pairs formatted as A-B
          \item location.B: it records locations of B in gene pairs formatted as A-B
          \item type.A: it records molecular functions of A in gene pairs formatted as A-B
          \item type.B: it records molecular functions of B in gene pairs formatted as A-B
        }
}
}
\description{
`ExtractTargetOnepairClusters` returns one pair of interaction from user-given cluster names.
}
\details{
The direction of this function is C-D, corresponding to coordinates X-Y in plot, and 
the subset options applied before will be inherited, which means that options
applied on X-axis clusters will be also on C, so as options on Y-axis will be on D.
}
