% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FastAlignPairs}
\alias{FastAlignPairs}
\title{Fast Align Gene Pairs}
\usage{
FastAlignPairs(
  xxpairs,
  ind.colname.end.dual,
  use.cols = c(1, 2),
  use.class = "integer"
)
}
\arguments{
\item{xxpairs}{Gene pairs to be compared.}

\item{ind.colname.end.dual}{Gene Pairs are given as pairs, so as their properties.}

\item{use.cols}{Use which columns to align those pairs. Column ID or column name is supported.}

\item{use.class}{Defines what is given in \code{use.cols}.}
}
\description{
This function helps to align gene pairs by their GeneID or GeneName. Its purpose is
to make different gene pairs data comparable in the same settings.
}
