% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-prep.R
\name{FetchInterOI}
\alias{FetchInterOI}
\title{Fetch One Interaction}
\usage{
FetchInterOI(object, cluster.x, cluster.y, if.ignore.location = FALSE)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{cluster.x}{Name of one cluster.}

\item{cluster.y}{Name of the other cluster.}

\item{if.ignore.location}{If set FALSE, genes will be extended by their locations, i.e. one gene 
along with several locations.
If set TRUE, only one subcellular location for one specific gene will be reserved, 
but the locations reserved are randomly picked.}
}
\value{
A \code{InterCell} object.
}
\description{
This function is to fetch result of one interaction (all gene pairs between 2 cell clusters)
from network analysis, and generates standard data for downstream analysis. Any \code{NA} about gene itself
(GeneID, GeneName) will be removed.
}
\seealso{
'GetOneInter' for extracting gene pairs with NA reserved
}
