% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-find-specialgenes.R
\name{FindCommonChangedGenesInOneCluster}
\alias{FindCommonChangedGenesInOneCluster}
\title{Find common changed genes in one cluster}
\usage{
FindCommonChangedGenesInOneCluster(
  interact.pairs.acted,
  cluster.name,
  cluster.sender.or.receiver = TRUE,
  common.change.confidence = 0.8,
  cmp.extend.positive = 0.25,
  cmp.extend.negative = 0.25,
  ...
)
}
\arguments{
\item{interact.pairs.acted}{TODO}

\item{cluster.name}{TODO}

\item{cluster.sender.or.receiver}{TODO}

\item{common.change.confidence}{TODO}

\item{cmp.extend.positive}{TODO}

\item{cmp.extend.negative}{TODO}

\item{...}{[TODO] a lot of param in \code{FindSpecialGenesInOnepairCluster} will be passed}
}
\value{
TODO
}
\description{
TODO
}
\details{
TODO
}
