% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{FormatCustomGenePairs}
\alias{FormatCustomGenePairs}
\title{Format User Given Gene Pairs}
\usage{
FormatCustomGenePairs(gene.pairs.table, species)
}
\arguments{
\item{gene.pairs.table}{[TODO] in data.frame 2 columns}

\item{species}{[TODO] which species to be add database}
}
\value{
A list with \code{$result} storing the formatted gene pairs.
}
\description{
This function makes 2-column table (each row represents one gene pair) to be 
formatted to be used in this package.
The gene names will be remapped to the authorized gene names recorded in gene 
reference database embedded in this package.
}
