% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{FullView.formula.Evaluation.default}
\alias{FullView.formula.Evaluation.default}
\title{Formula Used in Network Analysis}
\usage{
FullView.formula.Evaluation.default(
  pairs.ref,
  colname.eval = c("inter.LogFC.A", "inter.LogFC.B")
)
}
\arguments{
\item{pairs.ref}{data.frame. Data recording all gene pairs as well as their attributes in one interaction.}

\item{colname.eval}{character. The column names to be used to evaluate strength.}
}
\description{
This is the default formula for calculating overall strength for interactions.
}
