% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{setGenePairSplit}
\alias{setGenePairSplit}
\alias{setGenePairSplit,InterCell-method}
\alias{getGenePairSplit}
\alias{getGenePairSplit,InterCell-method}
\title{Using Gene Pair Split}
\usage{
setGenePairSplit(object, ...)

\S4method{setGenePairSplit}{InterCell}(object, new.gene.pair.split)

getGenePairSplit(object)

\S4method{getGenePairSplit}{InterCell}(object)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{...}{Parameters passed to other methods.}

\item{new.gene.pair.split}{A new gene pair split, which is either one character or string.}
}
\description{
The \code{setGenePairSplit} function is to \bold{set} the gene pair split, 
which is either one charater or string, and used to split 2 gene partners in one gene pair.

The \code{getGenePairSplit} function is to \bold{get} the gene pair split, 
which is either one charater or string, and used to split 2 gene partners in one gene pair.
}
\examples{
\dontrun{
  setGenePairSplit(object, some.new.gene.pair.split)
}

\dontrun{
  getGenePairSplit(object)
}
}
