% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes.human.ref.db}
\alias{genes.human.ref.db}
\title{Reference database of human genes}
\format{
An object of class \code{list} of length 3.
}
\source{
<ftp://ftp.ncbi.nih.gov/gene/>, and file path is /DATA/GENE_INFO/Mammalia/Homo_sapiens.gene_info.gz.
}
\usage{
genes.human.ref.db
}
\description{
A dataset contains the basic informations of 41745 human genes.

format List.
\itemize{
   \item {\code{gene.ncbi.db}: } {it is downloaded from NCBI ftp site:"/DATA/GENE_INFO/Mammalia/Homo_sapiens.gene_info.gz".}
   \item {\code{gene.synonyms.db}: } {data.frame, it records the synonyms-genenames pairs, which are generated from \code{gene.ncbi.db}.
         It maps gene synonyms to its authorized name and ID.
         }
   \item {\code{gene.dup.synonyms.db}: } {data.frame, which documents synonyms with multiple matching genes.}
}
}
\keyword{datasets}
