% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes.mouse.ref.db}
\alias{genes.mouse.ref.db}
\title{Reference database of mouse genes}
\format{
List.
\itemize{
   \item {\code{gene.ncbi.db}: } {it is downloaded from NCBI ftp site:"/DATA/GENE_INFO/Mammalia/Mus_musculus.gene_info.gz".}
   \item {\code{gene.synonyms.db}: } {data.frame, it records the synonyms-genenames pairs, which are generated from \code{gene.ncbi.db}.
         It maps gene synonyms to its authorized name and ID.
         }
   \item {\code{gene.dup.synonyms.db}: } {data.frame, which documents synonyms with multiple matching genes.}
}
}
\source{
<ftp://ftp.ncbi.nih.gov/gene/>, and file path is /DATA/GENE_INFO/Mammalia/Mus_musculus.gene_info.gz.
}
\usage{
genes.mouse.ref.db
}
\description{
A dataset contains the basic informations of 69864 mouse genes.
}
\keyword{datasets}
