% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-prep.R
\name{GetOneInter}
\alias{GetOneInter}
\title{Get One Interaction}
\usage{
GetOneInter(object, cluster.x, cluster.y)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{cluster.x}{Name of one cluster.}

\item{cluster.y}{Name of the other cluster.}
}
\value{
A list of all results for one interaction.
}
\description{
This function is to get the result of one interaction (all gene pairs between 2 cell clusters)
from network analysis, and directly return the value (even NA appears).
}
