% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-part1.R
\name{GetResult.PieActionEffect}
\alias{GetResult.PieActionEffect}
\title{Give composition proportion of action effect}
\usage{
GetResult.PieActionEffect(
  onepair.dgsa,
  limits.exprs.change = c("Xup.Yup", "Xup.Ydn", "Xdn.Yup", "Xdn.Ydn"),
  limits.action.effect = c("A-->B", "A<--B", "A--|B", "A|--B", "A--oB", "Ao--B",
    "A---B"),
  color.palette = c("#FB8072", "#B3DE69", "#80B1D3", "#8DD3C7", "#FFFFB3", "#BEBADA",
    "#FDB462"),
  legend.title = "Action Effect",
  legend.title.size = element_text(size = 12),
  legend.key.size = unit(6, "mm"),
  legend.text.size = element_text(size = 8),
  legend.box.margin = margin(0, 0, 0, 6),
  caption.label.size = 12,
  caption.label.x = 0.5,
  caption.label.y = 1,
  caption.label.hjust = 0.5,
  caption.label.vjust = 1
)
}
\arguments{
\item{onepair.dgsa}{List. Return value of \code{\link{CollectHierachyOnepairClusters}}.}

\item{limits.exprs.change}{Character. Use exprssion level change of clusters to select part of data to be shown.}

\item{limits.action.effect}{Character. Select some action effects to be put in the result.}

\item{color.palette}{It gives the colours for plotting.}

\item{legend.title}{It sets the attributes of legend title, and should be \code{element_text()}.}

\item{legend.key.size}{The size of keys in legend. It should be in unit format, see \code{?unit} for further help.}

\item{legend.box.margin}{It sets the margin of legend box, and should be \code{margin()}, see \code{?margin} for further help.}

\item{caption.label.size}{It sets the size of caption label.}

\item{caption.label.x}{It sets the position of caption label in x-axis.}

\item{caption.label.y}{It sets the position of caption label in y-axis.}

\item{caption.label.hjust}{It sets the alignment type of caption label in horizontal level, see \code{vignette("ggplot2-specs")}
for further help.}

\item{caption.label.vjust}{It sets the alignment type of caption label in vertical level, see \code{vignette("ggplot2-specs")}
for further help.}

\item{legend.text}{It sets the attributes of legend annotation texts, and should be \code{element_text()}.}
}
\value{
A list. Use \code{Tool.ShowGraph()} to see the \bold{plot}, \code{Tool.WriteTables()} to save the result \bold{table} in .csv files.
\itemize{
  \item plot: the object of \pkg{ggplot2}.
  \item table: a list of \code{data.frame}.
}
}
\description{
This function focuses on one interaction pair and its detailed information(expression changes, 
gene-gene action effects), and get result from it.
}
\examples{
# select only Xup.Yup
\dontrun{
GetResult.SummaryOnepairClusters(
  onepair.gmoc,
  limits.exprs.change = c("Xup.Yup")
)
}



}
