% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed-view-in-targets.R
\name{GetResult.PlotOnepairClusters.CellPlot}
\alias{GetResult.PlotOnepairClusters.CellPlot}
\title{Draw CellPlot for one pair of clusters}
\usage{
GetResult.PlotOnepairClusters.CellPlot(
  VEinfos,
  area.extend.times = 1,
  hide.locations.A = NULL,
  hide.types.A = NULL,
  hide.locations.B = NULL,
  hide.types.B = NULL,
  hide.sole.vertices = TRUE,
  nodes.size = 3,
  nodes.colour = "grey",
  nodes.alpha = 0.8,
  nodes.shape = 21,
  nodes.stroke = 0.7,
  nodes.label.size = 2,
  nodes.label.colour = "white",
  link.size = 0.1,
  link.colour = c("red", "green", "blue", "lightgrey"),
  link.alpha = 0.7,
  link.arrow.angle = 20,
  link.arrow.length = 10,
  link.arrow.type = "open"
)
}
\arguments{
\item{VEinfos}{List. It contains informations about vertices and edges, and is exactly return value of
\code{GenerateVEinfos()}.}

\item{area.extend.times}{Numeric. If a warning given like "Cannot be located inside!" or something else, 
one should change this paramemter to be larger to get all vertices allocated.}

\item{hide.locations.A}{Character. It applies extra limitation on the locations of A in gene pairs formatted as A-B.}

\item{hide.types.A}{Character. It applies extra limitation on the types(molecular functions) of A in gene pairs formatted as A-B.}

\item{hide.locations.B}{Character. It applies extra limitation on the locations of B in gene pairs formatted as A-B.}

\item{hide.types.B}{Character. It applies extra limitation on the types(molecular functions) of B in gene pairs formatted as A-B.}

\item{hide.sole.vertices}{Character. It hides sole vertices which have no available edges.}

\item{nodes.size}{Numeric. Size of nodes.}

\item{nodes.colour}{Character. Colour of nodes.}

\item{nodes.alpha}{Numeric. Alpha of nodes.}

\item{nodes.shape}{Vector. Use shape reprentable IDs(in integer) or directly shape description(in character). 
See \pkg{ggplot2} for details}

\item{nodes.stroke}{Numeric. Stroke size of nodes. See details in \pkg{ggplot2}.}

\item{nodes.label.size}{Numeric. Size of label on nodes.}

\item{nodes.label.colour}{Character. Colour of label on nodes.}

\item{link.size}{Numeric. Size of link width.}

\item{link.colour}{Character. Colour of links, the length should be same as \code{CellTalkDB::kpred.action.effect}.}

\item{link.alpha}{Numeric. Alpha of link.}

\item{link.arrow.angle}{Numeric. Angle of link arrow.}

\item{link.arrow.length}{Numeric. Length of link arrow.}

\item{link.arrow.type}{Character. Type of link arrow, either \code{open} or \code{closed}.}
}
\value{
List. Use \code{Tool.ShowPlot()} to see the \bold{plot}, \code{Tool.WriteTables()} to save the result \bold{table} in .csv files.
\itemize{
  \item plot: the object of \pkg{ggplot2}.
  \item table: a list of \code{data.frame}.
}
}
\description{
This function analyzes the interaction pairs between two target clusters, and gives one vivid two-cell
graph and several result tables in the return values.
}
