% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed-view-part3-sub1.R
\name{GetResult.PlotOnepairClusters.CellPlot.SmallData}
\alias{GetResult.PlotOnepairClusters.CellPlot.SmallData}
\title{Draw CellPlot for one pair of clusters}
\usage{
GetResult.PlotOnepairClusters.CellPlot.SmallData(
  VEinfos,
  area.extend.times = 10,
  hide.locations.A = c("Other"),
  hide.types.A = NULL,
  hide.locations.B = c("Other"),
  hide.types.B = NULL,
  hide.sole.vertices = FALSE,
  expand.gap.radius.list = list(ECM = 2, CTP = 2, NC = 2, OTHER = 2),
  expand.shift.degree.list = list(ECM = 90, CTP = 30, NC = 30, OTHER = 30),
  expand.gap.degree.list = list(ECM = 180, CTP = 60, NC = 60, OTHER = 60),
  expand.center.density.list = list(ECM = 0.25, NC = 0.25, OTHER = 0.25),
  expand.outside.cut.percent.list = list(ECM = 0.03, NC = 0.03, OTHER = 0.03),
  expand.PM.gap.len = 2,
  locate.PM.method = "uniform",
  nodes.size.range = c(6, 12),
  nodes.size.gap = 1,
  nodes.fill.updn = c("#D07F86", "#7AAF7A"),
  nodes.colour = c("lightgrey"),
  nodes.alpha = 1,
  nodes.shape = 21,
  nodes.stroke = 0,
  label.size.nodes = c(3, 3),
  label.colour.nodes = c("black", "black"),
  label.vjust = c(0, 0),
  label.hjust = c(1, 0),
  label.nudge.x = c(0, 0),
  label.nudge.y = c(2, 2),
  label.padding.itself = list(unit(0.25, "lines"), unit(0.25, "lines")),
  label.size.itself = c(0.25, 0.25),
  link.size = 0.6,
  link.colour = c("#D70051", "#00913A", "#956134", "#B5B5B6"),
  link.alpha = 1,
  link.arrow.angle = 20,
  link.arrow.length = 10,
  link.arrow.type = "open",
  legend.show.fill.updn.label = c("UP", "DN"),
  legend.show.fill.override.point.size = 3,
  legend.show.size.override.colour = "black",
  legend.show.size.override.stroke = 0.7,
  legend.show.size.override.size.proportion = 1
)
}
\arguments{
\item{VEinfos}{List. It contains informations about vertices and edges, and is exactly return value of
\code{GenerateVEinfos()} or \code{TrimVEinfos()}.}

\item{area.extend.times}{Numeric. Its default value is 10 which can handle most cases. If a warning given like "Cannot be located inside!" or something else, 
one should change this paramemter to be larger to get all vertices allocated.}

\item{hide.locations.A}{Character. It applies extra limitation on the locations of A in gene pairs formatted as A-B.}

\item{hide.types.A}{Character. It applies extra limitation on the types(molecular functions) of A in gene pairs formatted as A-B.}

\item{hide.locations.B}{Character. It applies extra limitation on the locations of B in gene pairs formatted as A-B.}

\item{hide.types.B}{Character. It applies extra limitation on the types(molecular functions) of B in gene pairs formatted as A-B.}

\item{hide.sole.vertices}{Character. It hides sole vertices which have no available edges.}

\item{expand.gap.radius.list}{Numeric. It defines the minimum distance between points(genes) for each plotting area.}

\item{expand.shift.degree.list}{Numeric. It defines the begin degree that points(genes) are to be drawn. The degree is calculated counter-clockwise.}

\item{expand.gap.degree.list}{Numeric. It defines the way that points(genes) arrange. If it is set 180 and shift degree is set 90, then points will be aligned in vertical line. 
If it is set 90 and shift degree is set 0, then points will be put counter-clockwise from horizontal line to vertical and then back to horizontal and finally at vertical place.}

\item{expand.center.density.list}{Numeric. It defines the density in each plotting area. Higher value means points concentrating more around 
the center of the area, and lower value means more sparse.}

\item{expand.outside.cut.percent.list}{[TODO]}

\item{expand.PM.gap.len}{[TODO]}

\item{locate.PM.method}{[TODO]}

\item{nodes.size.range}{[TODO]}

\item{nodes.size.gap}{[TODO]}

\item{nodes.fill.updn}{[TODO]}

\item{nodes.colour}{Character. Colour of nodes.}

\item{nodes.alpha}{Numeric. Alpha of nodes.}

\item{nodes.shape}{Vector. Use shape reprentable IDs(in integer) or directly shape description(in character). 
See \pkg{ggplot2} for details}

\item{nodes.stroke}{Numeric. Stroke size of nodes. See details in \pkg{ggplot2}.}

\item{label.size.nodes}{[TODO]}

\item{label.colour.nodes}{[TODO]}

\item{label.vjust}{[TODO]}

\item{label.hjust}{[TODO]}

\item{label.nudge.x}{[TODO]}

\item{label.nudge.y}{[TODO]}

\item{label.padding.itself}{[TODO]}

\item{label.size.itself}{[TODO]}

\item{link.size}{Numeric. Size of link width.}

\item{link.colour}{Character. Colour of links, the length should be same as \code{InterCellDB::kpred.action.effect}.}

\item{link.alpha}{Numeric. Alpha of link.}

\item{link.arrow.angle}{Numeric. Angle of link arrow.}

\item{link.arrow.length}{Numeric. Length of link arrow.}

\item{link.arrow.type}{Character. Type of link arrow, either \code{open} or \code{closed}.}

\item{legend.show.fill.updn.label}{[TODO]}

\item{legend.show.fill.override.point.size}{[TODO]}

\item{legend.show.size.override.colour}{[TODO]}

\item{legend.show.size.override.stroke}{[TODO]}
}
\value{
List. Use \code{Tool.ShowPlot()} to see the \bold{plot}, \code{Tool.WriteTables()} to save the result \bold{table} in .csv files.
\itemize{
  \item plot: the object of \pkg{ggplot2}.
  \item table: a list of \code{data.frame}.
}
}
\description{
This function analyzes the interaction pairs between two target clusters, and gives one vivid two-cell
graph and several result tables in the return values.
}
