% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-in-fullview.R
\name{GetResult.SummaryClustersInteracts}
\alias{GetResult.SummaryClustersInteracts}
\title{Get result of analysis in full view}
\usage{
GetResult.SummaryClustersInteracts(
  interact.pairs.acted,
  show.clusters.in.x = NULL,
  show.clusters.in.y = NULL,
  power.max.limit = NULL,
  power.min.limit = NULL,
  hide.power.label = FALSE,
  cnt.max.limit = NULL,
  cnt.min.limit = NULL,
  hide.cnt.label = FALSE,
  nodes.size = c(1, 6),
  nodes.colour.seq = c("#00809D", "#EEEEEE", "#C30000"),
  nodes.colour.value.seq = c(0, 0.5, 1),
  label.power.options = list(hjust = "middle", vjust = "top", nudge.x = 0, nudge.y =
    -0.3, size = 2),
  label.cnt.options = list(hjust = "left", vjust = "center", nudge.x = 0.3, nudge.y =
    0, size = 2),
  plot.axis.x.name = "clusters-x",
  plot.axis.y.name = "clusters-y"
)
}
\arguments{
\item{interact.pairs.acted}{List. The return value of \code{\link{AnalyzeClustersInteracts}}, 
which is formatted data about target interaction pairs.}

\item{show.clusters.in.x}{Vector. Clusters(use cluster names) that are chosen to show in x-axis by users.}

\item{show.clusters.in.y}{Vector. Clusters(use cluster names) that are chosen to show in y-axis by users.}

\item{power.max.limit}{Numeric. Specify the upper limit of power, whose value is highly user-defined and data-dependent.}

\item{power.min.limit}{Numeric. Specify the lower limit of power, like \code{power.max.limit}.}

\item{hide.power.label}{Logic. If TRUE, the label appended for power value will be hidden, otherwise, the label will be kept.}

\item{cnt.max.limit}{Numeric. Specify the upper limit of count, whose value is highly user-defined and data-dependent.}

\item{cnt.min.limit}{Numeric. Specify the lower limit of count, like \code{cnt.max.limit}.}

\item{hide.cnt.label}{Logic. If TRUE, the label appended for count value will be hidden, otherwise, the label will be kept.}

\item{nodes.size}{[TODO]}

\item{nodes.colour.seq}{Character. Given colours will be used to generate colour gradient for plotting.}

\item{nodes.colour.value.seq}{Numeric. If set NULL, evenly place each colour in \code{nodes.colour.seq} vector.
Otherwise, numeric values with the same length of \code{nodes.colour.seq} should be given to specify the positions corresponding to each colour.
See parameter \code{values} in \code{ggplot2::scale_colour_gradientn} for details.}

\item{label.power.options}{List. Options are hjust, vjust, nudge.x, nudge.y, size. See \code{ggplot2::geom_text} for details.
\itemize{
  \item hjust:   horizontal justification. Use either a string in ("left", "center", "right"), or a number between 0 and 1. See
                 \code{vignette("ggplot2-specs", package = "ggplot2")} for details.
  \item vjust:   vertical justification. Use either a string in ("top", "middle", "bottom"), or a number between 0 and 1. See
                 \code{vignette("ggplot2-specs", package = "ggplot2")} for details.
  \item nudge.x: horizontal adjustment to nudge labels by.
  \item nudge.y: vertical adjustment to nudge labels by.
  \item size:    label size.
}}

\item{label.cnt.options}{List. Options are like \code{label.power.options}.}

\item{plot.axis.x.name}{Character. X-axis name when plot is shown.}

\item{plot.axis.y.name}{Character. Y-axis name when plot is shown.}
}
\value{
List. Use \code{Tool.ShowPlot()} to see the \bold{plot}, \code{Tool.WriteTables()} to save the result \bold{table} in .csv files.
\itemize{
  \item plot: the object of \pkg{ggplot2}.
  \item table: a list of \code{data.frame}.
}
}
\description{
This function focuses on some subset of interaction pairs, and get rather fine grained
result from those.
}
\details{
This function uses some subset of interaction pairs, and does calculations in cluster level.
It calculates:
\itemize{
  \item count: count of interaction pairs.
  \item power: strength of interaction pairs, which is calculated by formula: SUM(abs(LogFC[geneA] * LogFC[geneB])), 
               as the function defines: \code{FullView.Evaluation.func.default}.
}
}
\examples{
\dontrun{
# show interaction pairs in some clusters, and set power & cnt max.limit
GetResult.SummaryClustersInteracts(interact.pairs.acted, 
  show.clusters.in.x = c("Microglia", "Astrocyte"), 
  show.clusters.in.y = c("Endothelial cell"),
  power.max.limit = 1.6,
  cnt.max.limit = 50
)
  
# change the label
GetResult.SummaryClustersInteracts(interact.pairs.acted, 
  show.clusters.in.x = c("Astrocyte"), 
  show.clusters.in.y = NULL,  # show all clusters in y-axis
  power.max.limit = 1.6,
  cnt.max.limit = 30,
  label.power.options = list(nudge.y = -0.2, size = 3),
  label.cnt.options = list(hjust = "left", vjust = "center", nudge.x = 0.4, nudge.y = 0, size = 2)
)
}



}
