% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-find-specialgenes.R
\name{GetResult.SummarySpecialGenes}
\alias{GetResult.SummarySpecialGenes}
\title{Summary special genes in cluster groups}
\usage{
GetResult.SummarySpecialGenes(
  onepair.spgenes,
  VEinfos,
  show.uq.cnt.range = c(1),
  show.uq.cnt.merged = TRUE,
  select.genepairs = character(),
  select.genepairs.method = "random",
  select.by.method.pairs.limit = 10,
  show.topn.inside.onepair = 2,
  show.cluster.group.order = character(),
  plot.font.size.base = 12,
  facet.scales = "free_x",
  facet.space = "free_x",
  facet.text.x = element_text(size = 8, colour = "black"),
  facet.background = element_rect(fill = "lightgrey", colour = "white"),
  bar.colour = character(),
  bar.width = 0.8,
  axis.text.x.pattern = element_text(angle = 90, vjust = 0.5, hjust = 1)
)
}
\arguments{
\item{onepair.spgenes}{[TODO]}

\item{VEinfos}{List. It contains informations about vertices and edges, and is exactly return value of
\code{GenerateVEinfos()} or \code{TrimVEinfos()}.}

\item{show.uq.cnt.range}{[TODO]}

\item{show.uq.cnt.merged}{[TODO]}

\item{select.genepairs}{[TODO]}

\item{select.genepairs.method}{[TODO]}

\item{select.by.method.pairs.limit}{[TODO]}

\item{show.topn.inside.onepair}{[TODO]}

\item{show.cluster.group.order}{[TODO]}

\item{plot.font.size.base}{[TODO]}

\item{facet.scales}{[TODO]}

\item{facet.space}{[TODO]}

\item{facet.text.x}{[TODO]}

\item{facet.background}{[TODO]}

\item{bar.colour}{[TODO]}

\item{bar.width}{[TODO]}

\item{axis.text.x.pattern}{[TODO]}
}
\value{
List. [TODO]
}
\description{
[TODO]
}
\details{
[TODO]
}
