% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullView.R
\name{GetResultFullView}
\alias{GetResultFullView}
\title{Get Result for Network Analysis}
\usage{
GetResultFullView(
  object,
  show.clusters.in.x = NULL,
  show.clusters.in.y = NULL,
  sel.cluster.group.method = "all",
  power.apply.func = NULL,
  cnt.apply.func = NULL,
  plot.power.range = NULL,
  plot.cnt.range = NULL,
  nodes.size.range = c(1, 6),
  nodes.colour.seq = c("#00809D", "#EEEEEE", "#C30000"),
  nodes.colour.value.seq = c(0, 0.5, 1),
  label.power.options = list(hjust = "middle", vjust = "top", nudge.x = 0, nudge.y =
    -0.3, size = 2),
  label.cnt.options = list(hjust = "left", vjust = "center", nudge.x = 0.3, nudge.y =
    0, size = 2),
  plot.axis.x.name = "clusters-x",
  plot.axis.y.name = "clusters-y"
)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{show.clusters.in.x}{Clusters chosen to show in x-axis, corresponding to paramter \code{sel.clusters.X} in \code{\link{AnalyzeInterInFullView}}.}

\item{show.clusters.in.y}{Clusters chosen to show in y-axis, corresponding to paramter \code{sel.clusters.Y} in \code{\link{AnalyzeInterInFullView}}.}

\item{sel.cluster.group.method}{Options are 'all', 'diagonal' and 'diagonal-2'. It controls 
'all' is recommended for most cases, and 
'diagonal*' is recommended when same restrictions are applied to clusters in X and Y.}

\item{power.apply.func}{The function used to modify the power value, usually \code{log1p}, etc.}

\item{cnt.apply.func}{The function used to modify the count value, usually \code{sqrt} to avoid extremes problem.}

\item{plot.power.range}{This will extend the plotting range beyond actual value range for power.}

\item{plot.cnt.range}{This will extend the plotting range beyond actual value range for count.}

\item{nodes.size.range}{2 numbers that specifies the range of sizes of the nodes shown in the graph.}

\item{nodes.colour.seq}{Given colours will be used to generate colour gradient for plotting.}

\item{nodes.colour.value.seq}{Numeric. If set NULL, colours given in \code{nodes.colour.seq} will be evenly placed.
Otherwise, numeric values with the same length of \code{nodes.colour.seq} should be given to specify the positions corresponding to each colour.
See parameter \code{values} in \code{ggplot2::scale_colour_gradientn} for details.}

\item{label.power.options}{List of 'power' label appearance control parameters, including hjust, vjust, nudge.x, nudge.y, size. See details for help.}

\item{label.cnt.options}{List of label appearance control parameters like those in \code{label.power.options}.}

\item{plot.axis.x.name}{X-axis name when plotting.}

\item{plot.axis.y.name}{Y-axis name when plotting.}
}
\value{
List. Use \code{Tool.ShowPlot()} to see the \bold{plot}, \code{Tool.WriteTables()} to save the result \bold{table} in .csv files.
\itemize{
  \item plot: the object of \pkg{ggplot2}.
  \item table: a list of \code{data.frame}.
}
}
\description{
This function summarizes the result of network analysis in graph and table, and gives clues to 
most active intercellular communications.
}
\details{
\code{label.power.options} and \code{label.cnt.options}: 
The meanings for each item are recommended to refer to \code{ggplot2::geom_text} for details.
\itemize{
  \item hjust:   horizontal justification. Use either a string in ("left", "center", "right"), or a number between 0 and 1. See
                 \code{vignette("ggplot2-specs", package = "ggplot2")} for details.
  \item vjust:   vertical justification. Use either a string in ("top", "middle", "bottom"), or a number between 0 and 1. See
                 \code{vignette("ggplot2-specs", package = "ggplot2")} for details.
  \item nudge.x: horizontal adjustment to nudge labels by.
  \item nudge.y: vertical adjustment to nudge labels by.
  \item size:    label size.
}
}
