% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-part1.R
\name{GetResultPieActionMode}
\alias{GetResultPieActionMode}
\title{Give composition proportion of action mode}
\usage{
GetResultPieActionMode(
  object,
  limits.exprs.change = kexprs.change,
  limits.action.mode = kpred.mode,
  color.action.mode = kpred.color.mode,
  legend.title = "Action Mode",
  legend.title.size = element_text(size = 14),
  legend.key.size = unit(7, "mm"),
  legend.text.size = element_text(size = 12),
  legend.box.margin = margin(0, 0, 0, 6),
  show.note = TRUE,
  note.label.size = 3,
  note.lineheight = 1,
  note.postion.xy = c(0.5, 0.5),
  note.hjust = 0.5,
  note.vjust = 0.5
)
}
\arguments{
\item{object}{[TODO]}

\item{limits.exprs.change}{Character. Use exprssion level change of clusters to select part of data to be shown.}

\item{limits.action.mode}{Character. Select some action effects to be put in the result.}

\item{legend.title}{It sets the content of legend title.}

\item{legend.title.size}{It sets size of legend title.}

\item{legend.key.size}{The size of keys in legend.}

\item{legend.text.size}{It sets size of legend text.}

\item{legend.box.margin}{It sets the margin of legend box, and should be \code{margin()}, see \code{?margin} for further help.}

\item{show.note}{[TODO]}

\item{note.label.size}{[TODO]}

\item{note.lineheight}{[TODO]}

\item{note.postion.xy}{[TODO]}

\item{note.hjust}{[TODO]}

\item{note.vjust}{[TODO]}
}
\description{
This function focuses on one interaction pair and its detailed information(expression changes, 
gene-gene action mode), and get result from it.
}
