% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-part4.R
\name{GetResultTgCellPlot}
\alias{GetResultTgCellPlot}
\title{Get Result for Sptial Pattern}
\usage{
GetResultTgCellPlot(
  object,
  area.extend.times = 10,
  hide.other.area = TRUE,
  hide.locations.X = NULL,
  hide.locations.Y = NULL,
  hide.some.genes.X = NULL,
  hide.some.genes.Y = NULL,
  hide.sole.vertices = TRUE,
  expand.gap.radius.list = list(ECM = 2, CTP = 2, NC = 2, OTHER = 2),
  expand.shift.degree.list = list(ECM = 90, CTP = 30, NC = 30, OTHER = 30),
  expand.gap.degree.list = list(ECM = 180, CTP = 60, NC = 60, OTHER = 60),
  expand.center.density.list = list(ECM = 0.25, NC = 0.25, OTHER = 0.25),
  expand.outside.cut.percent.list = list(ECM = 0.03, NC = 0.03, OTHER = 0.03),
  expand.PM.gap.len = 2,
  locate.PM.method = "uniform",
  nodes.size.range = c(6, 12),
  nodes.size.gap = 1,
  nodes.fill.updn = c("#D07F86", "#7AAF7A"),
  nodes.colour = c("lightgrey"),
  nodes.alpha = 1,
  nodes.shape = 21,
  nodes.stroke = 0,
  label.size.nodes = c(3, 3),
  label.colour.nodes = c("black", "black"),
  label.vjust = c(0, 0),
  label.hjust = c(1, 0),
  label.nudge.x = c(0, 0),
  label.nudge.y = c(2, 2),
  label.padding.itself = list(unit(0.25, "lines"), unit(0.25, "lines")),
  label.size.itself = c(0.25, 0.25),
  link.size = 0.6,
  link.colour = c("#D70051", "#00913A", "#1296D4", "#956134", "#F46D42", "#0A0AFF",
    "#762A83", "#B5B5B6"),
  link.alpha = 1,
  link.linetype = c("solid", "solid", "solid", "44"),
  link.arrow.angle = c(20, 90, 60, 0),
  link.arrow.length = c(12, 6, 8, 0),
  link.arrow.type = c("closed", "open", "open", "open"),
  legend.show.fill.updn.label = c("UP", "DN"),
  legend.show.fill.override.point.size = 3,
  legend.show.size.override.colour = "black",
  legend.show.size.override.stroke = 0.7,
  legend.show.size.override.size.proportion = 1,
  legend.manual.left.spacing = unit(0.2, "cm"),
  legend.manual.internal.spacing = unit(0.6, "cm")
)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{area.extend.times}{Its default value is 10, which can handle most cases. If too many genes are collected,
users should enlarge this value until those genes can be allocated.}

\item{hide.other.area}{If set TRUE, genes with no subcellular location other than 'Other' will be removed from result 
as well as plotting area:Other.}

\item{hide.locations.X}{It removes genes (from cluster.X) in given subcellular locations from result.}

\item{hide.locations.Y}{It removes genes (from cluster.Y) in given subcellular locations from result.}

\item{hide.some.genes.X}{It removes genes (from cluster.X) from result.}

\item{hide.some.genes.Y}{It removes genes (from cluster.Y) from result.}

\item{hide.sole.vertices}{It decides whether to hide sole genes which have no available relation with other genes.}

\item{expand.gap.radius.list}{Numeric. It defines the minimum distance between points(genes) for each plotting area.}

\item{expand.shift.degree.list}{Numeric. It defines the begin degree that points(genes) are to be drawn. The degree is calculated counter-clockwise.}

\item{expand.gap.degree.list}{Numeric. It defines the way that points(genes) arrange. If it is set 180 and shift degree is set 90, then points will be aligned in vertical line. 
If it is set 90 and shift degree is set 0, then points will be put counter-clockwise from horizontal line to vertical and then back to horizontal and finally at vertical place.}

\item{expand.center.density.list}{Numeric. It defines the density in each plotting area. Higher value means points concentrating more around 
the center of the area, and lower value means more sparse.}

\item{expand.outside.cut.percent.list}{Numeric. It is used to restrict plotting area range against center point in each plotting area. The value of this parameter 
defines the width of outside not-drawing area, and if is set 0, means plotting over the total area, or if is set 0.5, means plotting inside circle of half radius.}

\item{expand.PM.gap.len}{It defines the gap of each points(genes) plotted in area:Plasma Membrane. The larger value means sparser the points being plotted, while smaller denser.}

\item{locate.PM.method}{It defines the way to allocate points(genes) plotted in area:Plasma Membrane. It supports 2 methods for now: 
'random' and 'uniform'.}

\item{nodes.size.range}{Numeric of length 2. The former gives the minimum size, while the latter gives the maximum. Node(genes) sizes are reflecting the actual LogFC value of every gene.}

\item{nodes.size.gap}{It is used along with \code{nodes.size.range}. It defines the resolution of changing point sizes. For example, defines \code{nodes.size.range = c(3,6)} and 
\code{nodes.size.gap = 1}, then point sizes are \code{c(3,4,5,6)}.}

\item{nodes.fill.updn}{Character of length 2. It defines the 2 colours used to distinguish the up-regulated and down-regulated nodes(genes).}

\item{nodes.colour}{Color of nodes.}

\item{nodes.alpha}{Alpha of nodes.}

\item{nodes.shape}{Use shape reprentable IDs(in integer) or directly shape description(in character). 
See \pkg{ggplot2} for details}

\item{nodes.stroke}{Stroke size of nodes. See details in \pkg{ggplot2}.}

\item{label.size.nodes}{It defines the size of label text on nodes(genes). If get length 2, the former defines size of nodes in the left part of graph, and 
the latter corresponds to the right part of graph.}

\item{label.colour.nodes}{It defines the colour of label.}

\item{label.vjust}{It defines the vertical alignment value for labels.}

\item{label.hjust}{It defines the horizontal alignment value for labels.}

\item{label.nudge.x}{It defines the slight adjustment movement along x-axis when plotting.}

\item{label.nudge.y}{It defines the slight adjustment movement along y-axis when plotting.}

\item{label.padding.itself}{It defines the padding size of label.}

\item{label.size.itself}{It defines the size of label itself(includes padding, etc).}

\item{link.size}{Size of link width.}

\item{link.colour}{Colour of links, the length should be same as \code{kpred.action.mode}.}

\item{link.alpha}{Alpha of links.}

\item{link.linetype}{Linetype of links.}

\item{link.arrow.angle}{Angle of link arrow.}

\item{link.arrow.length}{Length of link arrow.}

\item{link.arrow.type}{Type of link arrow, either \code{open} or \code{closed}.}

\item{legend.show.fill.updn.label}{Character of length 2. It gives the labels putting in legend:fill.}

\item{legend.show.fill.override.point.size}{Numeric. It changes the size of template points used in legend:fill.}

\item{legend.show.size.override.colour}{Character. It changes the colour of template points used in legend:colour.}

\item{legend.show.size.override.stroke}{Numeric. It changes the stroke of template points used in legend:colour.}

\item{legend.show.size.override.size.proportion}{Numeric. It changes the size of template points used in legend:size, to avoid 
unbalanced large or small size shown in legend.}

\item{legend.manual.left.spacing}{Unit. It spcifies the left spacing of the manual legend to \pkg{ggplot2} automatically-generated legends.}

\item{legend.manual.internal.spacing}{Unit. It defines the spacing between 2 manual legend about action mode and action effect.}
}
\value{
List. Use \code{Tool.ShowPlot()} to see the \bold{plot}, \code{Tool.WriteTables()} to save the result \bold{table} in .csv files.
\itemize{
  \item plot: the object of \pkg{ggplot2}.
  \item grid.plot: graphical output generated from \pkg{grid}.
  \item table: a list of \code{data.frame}.
}
}
\description{
This function collects gene pairs from one interaction (between 2 target clusters), and 
related gene attributes (subcellular location, expression), action properties (action mode and action effect).
Then, all those informations are summarized.
}
\details{
All parameters starts with 'label.' can be specified of 2 values or only 1 value. As the graph contains 2 cells (drawn like it), with 1 cell put in the 
left side and the other put in right side, there are some situations that number of nodes(genes) is quite unbalanced between them. The label may hard to 
be placed properly. To deal with this problem, this function gives the way to specify different label patterns for each part. That's why parameters started with 
'label.' can be set of 2 values as well as 1 value.
}
