% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-part2.R
\name{GetResultTgCrosstalk}
\alias{GetResultTgCrosstalk}
\title{Evaluation gene pairs for one pair of clusters}
\usage{
GetResultTgCrosstalk(
  object,
  plot.X.to.Y = TRUE,
  colnames.to.cmp = c("LogFC", "PVal"),
  range.to.use = list(LogFC = c(-Inf, Inf), PVal = c(-Inf, Inf)),
  axis.order.xy = c("AlphaBet", "AlphaBet"),
  axis.order.xy.decreasing = c(TRUE, TRUE),
  plot.font.size.base = 12,
  nodes.colour.seq = c("#00809D", "#EEEEEE", "#C30000"),
  nodes.colour.value.seq = c(0, 0.5, 1),
  nodes.size.range = c(2, 8),
  axis.text.x.pattern = element_text(angle = 30, hjust = 1)
)
}
\arguments{
\item{object}{[TODO]}

\item{plot.X.to.Y}{[TODO] if TRUE, X will be original X in fullview , Y will be original Y in fullview.}

\item{colnames.to.cmp}{Character. The colnames to be used as evaluation params, currently only 2 params are supported.
The 1st one will be plotted differently by different size of nodes, and 2nd one will be different by colour of nodes.}

\item{range.to.use}{List. It specifies the user specified ranges for evaluation params.}

\item{axis.order.xy}{Character. It determines how the gene names will be ordered in the axis when plotting.}

\item{axis.order.xy.decreasing}{Logic. It determines whether the orders are of decreasing pattern or increasing pattern.}

\item{plot.font.size.base}{Numeric. It defines the font size of texts such as labels and titles.}

\item{nodes.colour.seq}{Character. It specifies the colour sequence of the nodes.}

\item{nodes.colour.value.seq}{Numeric. It is along with the param \code{nodes.colour.seq}, and changes the colour expansion.}

\item{nodes.size.range}{Numeric. It specifies the size range of the nodes.}

\item{axis.text.x.pattern}{It defines the axis text style in x-axis.}
}
\value{
List. Use \code{Tool.ShowPlot()} to see the \bold{plot}, \code{Tool.WriteTables()} to save the result \bold{table} in .csv files.
\itemize{
  \item plot: the object of \pkg{ggplot2}.
  \item table: a list of \code{data.frame}.
}
}
\description{
This function evaluates the importance of some specific gene pairs of one pair of clusters, 
and the default evaluation params are LogFC and PValAdj.
}
\details{
This function uses user-selected gene pairs, and uses evalution params to calculate the relative importance of each gene pair.
It calculates in default settings:
\itemize{
  \item LogFC: the log of fold changes, which indicates the relative gene expression.
  \item PValAdj: the confidence of discovering the gene as differently expressed genes. In Seurat, it uses bonferroni correction.
}
}
