% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-part3.R
\name{GetResultTgSpecificity}
\alias{GetResultTgSpecificity}
\title{Get Result for Specificity Analysis}
\usage{
GetResultTgSpecificity(
  object,
  sel.uq.cnt.options = integer(),
  sel.gene.pairs = NULL,
  sel.gene.pairs.method = NULL,
  sel.by.method.count = 10,
  sel.by.method.decreasing = TRUE,
  prioritize.cluster.groups = character(),
  plot.name.X.to.Y = TRUE,
  plot.uq.cnt.merged = TRUE,
  bound.to.use = list(Power = c(-Inf, +Inf), Confidence = c(-Inf, +Inf)),
  func.to.use = list(Power = function(x) {     x }, Confidence = function(x) {     1/(1
    + x) }),
  grid.plot.ncol = 1,
  barplot.or.dotplot = FALSE,
  plot.font.size.base = 12,
  axis.text.x.pattern = element_text(angle = 90, vjust = 0.5, hjust = 1),
  bar.facet.scales = "free_x",
  bar.facet.space = "free_x",
  bar.facet.text.x = element_text(size = 8, colour = "black"),
  bar.facet.background = element_rect(fill = "lightgrey", colour = "white"),
  bar.colour = character(),
  bar.width = 0.8,
  dot.colour.seq = c("#00809D", "#EEEEEE", "#C30000"),
  dot.colour.value.seq = c(0, 0.5, 1),
  dot.size.range = c(2, 8),
  dot.y.order.in.alphabet = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{sel.uq.cnt.options}{It defines the range of co-occurence count to be shown in result. If no option is given,
it will use all available co-occurence count.}

\item{sel.gene.pairs}{Directly specify the desired gene pairs. It should be given in standard table that is generated 
by \code{\link{FormatCustomGenePairs}}. The order of gene pairs will be implicitly used for result graph and table.
In addition, it's strictly aligned to clusters, which is illustrated in \code{\link{AnalyzeInterInFullView}}.}

\item{sel.gene.pairs.method}{Options are: "random", "by-power" or not to use (by set NULL). It works only when no specific gene pairs are given in 
parameter \code{sel.gene.pairs}. Method "random" will randomly pick up some gene pairs. Method "by-power" will calculate the power of 
the 2 genes in every gene pairs, and order them by the calculated values. 
The order of gene pairs generated by method will be implicitly used for result graph and table.}

\item{sel.by.method.count}{It defines the maximum number of gene pairs to be fetched by any method.}

\item{sel.by.method.decreasing}{It works for method "by-power". If TRUE, result will be ordered in decreasing way, otherwise the increasing direction.}

\item{prioritize.cluster.groups}{It defines the most concerning cluster groups, and those cluster groups given in this parameter, will be 
finally plotted from the most left-side to right.}

\item{plot.name.X.to.Y}{If set FALSE, switch the position of 2 involving clusters in the original names of interaction. If set TRUE, keep still.}

\item{plot.uq.cnt.merged}{If set TRUE, then gene pairs of different count of co-occurence will be merged to show in the result, or gene pairs 
will be grouped by their count of co-occurence and be plotted accordingly.}

\item{bound.to.use}{It specifies the user specified bound for evaluation params. The values out of bound will be coerced to 
either lower bound or upper bound. Default no bound is set, i.e. [-Inf, +Inf]}

\item{func.to.use}{The function used to further transform the values, e.g. \code{log1p}.}

\item{grid.plot.ncol}{It gives the number of columns for plotting grid when \code{plot.uq.cnt.merged = FALSE}.}

\item{barplot.or.dotplot}{If TRUE, use 'barplot', or, use 'dotplot'.}

\item{plot.font.size.base}{It defines the base font size for all texts.}

\item{axis.text.x.pattern}{It defines the axis text style in x-axis.}

\item{bar.facet.scales}{It controls the scales that facet uses, and gets 4 options as defined by \pkg{ggplot2}: "fixed", "free", "free_x", "free_y".}

\item{bar.facet.space}{It controls the space allocating strategy that facet uses, and gets 4 options as defined by \pkg{ggplot2}: "fixed", "free", "free_x", "free_y".}

\item{bar.facet.text.x}{It defines the style of facet text on the top horizontal position.}

\item{bar.facet.background}{It defines the background style of labels among facets.}

\item{bar.colour}{It gives colors that plotting bars get to use. If no specific colour is given, then the 
built-in 20 kinds of colours will be used.}

\item{bar.width}{It defines the bar width.}

\item{dot.colour.seq}{It specifies the colour sequence for dots.}

\item{dot.colour.value.seq}{It is along with the param \code{dot.colour.seq}, and changes the colour expansion.}

\item{dot.size.range}{It specifies the size range of the dots}

\item{dot.y.order.in.alphabet}{If set TRUE, force to order gene pairs in alphabet,
which ignores the implicit order by \code{sel.gene.pairs} or \code{sel.gene.pairs.method}.}

\item{...}{Other parameter that can be passed to select by method functions.}
}
\value{
List. Use \code{Tool.ShowPlot()} to see the \bold{plot}, \code{Tool.WriteTables()} to save the result \bold{table} in .csv files.
\itemize{
  \item plot: the object of \pkg{ggplot2}.
  \item table: a list of \code{data.frame}.
}
}
\description{
This function is to summary gene pair specificity by showing all gene pairs with their 
co-occurring interactions. Gene pairs are evaluated on both power and confidence like it in \code{\link{GetResultTgCrosstalk}}.
}
