% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go.human.ref.db}
\alias{go.human.ref.db}
\title{GO reference database for human}
\format{
An object of class \code{data.frame} with 323030 rows and 9 columns.
}
\source{
<ftp://ftp.ncbi.nih.gov/gene/>, and file path is /DATA/gene2go.gz.
}
\usage{
go.human.ref.db
}
\description{
A dataset collects human genes and their GO annotations. It is extracted from
GO annotations of mixed species, and has 323030 rows in total.
}
\keyword{datasets}
