% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-pkg-param.R
\name{Inside.DummyActionsRefDB}
\alias{Inside.DummyActionsRefDB}
\title{Description of parameter actions.ref.db}
\usage{
Inside.DummyActionsRefDB(actions.ref.db)
}
\arguments{
\item{actions.ref.db}{Data.frame. Reference pairs actions database provided in this package, the list is given as follows:
\itemize{
  \item For human: \code{actions.human.ref.db}
  \item For mouse: \code{actions.mouse.ref.db}
}}
}
\description{
Use to provide the same parameter description for actions.ref.db
}
