% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkg-parameters.R
\name{Inside.DummyAnnoLocationRefDB}
\alias{Inside.DummyAnnoLocationRefDB}
\title{Description of parameter anno.location.ref}
\usage{
Inside.DummyAnnoLocationRefDB(anno.location.ref)
}
\arguments{
\item{anno.location.ref}{Data.frame. Annotation of subcellular locations for genes provided in this package, the list is given as follows:
\itemize{
  \item For human: \code{anno.location.human.ref.db}
  \item For mouse: \code{anno.location.mouse.ref.db}
}}
}
\description{
Use to provide the same parameter description for anno.location.ref
}
