% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkg-parameters.R
\name{Inside.DummyAnnoTypeRefDB}
\alias{Inside.DummyAnnoTypeRefDB}
\title{Description of parameter anno.type.ref}
\usage{
Inside.DummyAnnoTypeRefDB(anno.type.ref)
}
\arguments{
\item{anno.type.ref}{Data.frame. Annotation of types(molecular functions) for genes provided in this package, the list is given as follows:
\itemize{
  \item For human: \code{anno.type.human.ref.db}
  \item For mouse: \code{anno.type.mouse.ref.db}
}}
}
\description{
Use to provide the same parameter description for anno.type.ref
}
