% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-pkg-param.R
\name{Inside.DummyFgenes}
\alias{Inside.DummyFgenes}
\title{Description of parameter fgenes.remapped.all}
\usage{
Inside.DummyFgenes(fgenes.remapped.all)
}
\arguments{
\item{fgenes.remapped.all}{Data.frame. Informations about feature genes(markers), prefer using return value of 
\code{\link{DataPrep.RemapClustersMarkers}} and similar functions start with \code{DataPrep.} here.}
}
\description{
Use to provide the same parameter description for fgenes.remapped.all
}
