% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkg-parameters.R
\name{Inside.DummyGenesRefDB}
\alias{Inside.DummyGenesRefDB}
\title{Description of parameter genes.ref.db}
\usage{
Inside.DummyGenesRefDB(genes.ref.db)
}
\arguments{
\item{genes.ref.db}{List. It is taxonomy specific, which contains 3 parts:
1. gene.ncbi.db            NCBI-gene database for specific species
2. gene.synonyms.db        GeneID-GeneName-Synonym pairs extracted from (1)
3. gene.dup.synonyms.db  Some genes may have same synonyms, and this can cause mistakes 
when match target synonym back to its authorized genename.}
}
\description{
Use to provide the same parameter description for genes.ref.db
}
