% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-pkg-param.R
\name{Inside.DummyGORefDB}
\alias{Inside.DummyGORefDB}
\title{Description of parameter go.ref.db}
\usage{
Inside.DummyGORefDB(go.ref.db)
}
\arguments{
\item{go.ref.db}{Data.frame. Reference GO database provided in this package, the list is given as follows:
\itemize{
  \item For human: \code{go.human.ref.db}
  \item For mouse: \code{go.mouse.ref.db}
}}
}
\description{
Use to provide the same parameter description for go.ref.db
}
