% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkg-parameters.R
\name{Inside.DummyInteractPairsActed}
\alias{Inside.DummyInteractPairsActed}
\title{Description of parameter interact.pairs.acted}
\usage{
Inside.DummyInteractPairsActed(interact.pairs.acted)
}
\arguments{
\item{interact.pairs.acted}{List. The return value of \code{\link{AnalyzeClustersInteracts}}, 
which is formatted data about target interaction pairs.}
}
\description{
Use to provide the same parameter description for interact.pairs.acted
}
