% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkg-parameters.R
\name{Inside.DummyPairsRef}
\alias{Inside.DummyPairsRef}
\title{Description of parameter pairs.ref}
\usage{
Inside.DummyPairsRef(pairs.ref)
}
\arguments{
\item{pairs.ref}{Data.frame. Reference interaction database provided in this package, the list is given as follows: 
\itemize{
 \item For human: \code{pairs.human.experiments.db}, \code{pairs.human.knowledge.db}, \code{pairs.human.prediction.db}
 \item For mouse: \code{pairs.mouse.experiments.db}, \code{pairs.mouse.knowledge.db}, \code{pairs.mouse.prediction.db}
}}
}
\description{
Use to provide the same parameter description for pairs.ref
}
