% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkg-parameters.R
\name{Inside.DummyVEinfos}
\alias{Inside.DummyVEinfos}
\title{Description of parameter VEinfos}
\usage{
Inside.DummyVEinfos(VEinfos)
}
\arguments{
\item{VEinfos}{List. It contains informations about vertices and edges, and is exactly return value of
\code{GenerateVEinfos()} or \code{TrimVEinfos()}.}
}
\description{
Use to provide the same parameter description for VEinfos
}
