% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-pkg-param.R
\name{InsideObjectDBPack}
\alias{InsideObjectDBPack}
\title{Description of parameter object for InterCell-class}
\usage{
InsideObjectDBPack(db.object)
}
\arguments{
\item{db.object}{A \code{InterCellDBPack} object, which is created by \code{\link{CreateDBPackObject}}.}
}
\description{
Use to provide the same parameter description for object (InterCell-class).
}
