% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-pkg-param.R
\name{InsideObjectInterCell}
\alias{InsideObjectInterCell}
\title{Description of parameter object for InterCell-class}
\usage{
InsideObjectInterCell(object)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}
}
\description{
Use to provide the same parameter description for object (InterCell-class).
}
