% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-pkg-param.R
\name{InsideParam.species}
\alias{InsideParam.species}
\title{Description of parameter species}
\usage{
InsideParam.species(species)
}
\arguments{
\item{species}{\pkg{InterCellDB} supports species in either 'human' or 'mouse'.
Other species are not supported to be used for analysis.}
}
\description{
Use to provide the same parameter description for fgenes.remapped.all
}
