% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\docType{class}
\name{InterCell-class}
\alias{InterCell-class}
\alias{InterCell}
\title{The InterCell Class}
\description{
The InterCell object is used to in all the analysis \pkg{InterCellDB} provided.
It incoporates the necessary databases, and stores every important intermediate result.
}
\section{Slots}{

\describe{
\item{\code{fgenes}}{InterCellDB needs the differential expressed genes and their belonging clusters as input. 
The given data need to have columns named 'gene', 'cluster', 'LogFC', 'PVal'. Those each means, 'gene': the 
authorized gene names, 'cluster': cell cluster, 'LogFC': the log transformed fold change of genes, 'PVal': the 
P val for regarding one gene as differentially expressed one.}

\item{\code{database}}{Database stored in \code{InterCellDBPack} object.}

\item{\code{formulae}}{The default formulae to be used in analysis.}

\item{\code{inter.fullview}}{This is used to store the result of network analysis.}

\item{\code{tg.action.pairs}}{This is used to store the intermediate result of action properties and gene pairs.}

\item{\code{tg.veinfo}}{This is used to store the intermediate result of one intercellular communication (for one interacting 2-cell group).}

\item{\code{tg.action.comp}}{This is used to store the result of analyzing composition of action mode and action effect in one interacellular communication.}

\item{\code{tg.spgenes}}{This is used to store the selected gene pairs and all their participating interacitons in one intercellular communication.}

\item{\code{tool.vars}}{This stores several variables to be used embedded in program.}

\item{\code{misc}}{The not important intermediate result will be put in this parameter as well as some 
pre-defined settings.}
}}

