% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkg-parameters.R
\docType{package}
\name{InterCellDB-package}
\alias{InterCellDB}
\alias{InterCellDB-package}
\title{InterCellDB: Tools for Analyzing Interactions among Cells}
\description{
This tool provides databases covering human and mouse, and provides bunches of functions intended to serve for data analysis of scRNA-seq. Its databases collect a variety of data source, including STRING, NCBI-gene, COMPARTMENTS, GO, Uniprot, etc. Its function on analysis starts from fullview of cell-cell-interaction to specific interactions between specific clusters and intends to find a small subset of important gene pairs.
}
\details{
This package provides databases and tools for scRNA-seq data analysis.
}
\author{
\strong{Maintainer}: Jin Ziyang Developer \email{454785187@qq.com}

Authors:
\itemize{
  \item Shi Ligen Developer \email{slg0904@163.com} [reviewer]
  \item Zhang Xiaotao Developer \email{zhangxiaotao@zju.edu.cn} [data contributor]
}

}
\keyword{internal}
