% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\docType{class}
\name{InterCellDBPack-class}
\alias{InterCellDBPack-class}
\alias{InterCellDBPack}
\title{The Class for Database Storage in InterCellDB}
\description{
The InterCellDBPack class provides a standard storage format of all databases for running analysis. 
This object is mostly created along with \code{\link{InterCell-class}}. 
Users need to specify the species firstly. Then, user-defined subset of database could be 
collected by function \code{\link{SelectDBSubset}}.
}
\section{Slots}{

\describe{
\item{\code{species}}{InterCellDB only allows 'human' and 'mouse' species.}

\item{\code{genes.db}}{The gene reference database, which is generated from NCBI gene site. It records all the authorized gene names.}

\item{\code{pairs.db}}{The database records all gene pairs, which are transferred from protein interaction database - 'STRING'.}

\item{\code{actions.db}}{The database records all action properties (including 'action.mode', 'action.effect') for gene pairs in slot:`pairs.db`.}

\item{\code{anno.location.db}}{The database collects all subcellular location from COMPARTMENTS for genes in slot:`pairs.db`.}

\item{\code{anno.type.db}}{The database collects all molecular function from Uniprot for genes in slot:`pairs.db`.}

\item{\code{go.ref.db}}{The database collects all GO terms for genes in slot:`pairs.db`.}

\item{\code{accessory.db}}{A list of accessory database, for example, the mapping list of raw keywords from Uniport.}

\item{\code{misc}}{The preserved area for miscellaneous things. Currently no usage.}
}}

