% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalvariables.R
\docType{data}
\name{kClustersSplit}
\alias{kClustersSplit}
\title{Cluster split character (global).}
\format{
An object of class \code{character} of length 1.
}
\usage{
kClustersSplit
}
\description{
This is the character used to split clusternames in name of interaction pairs.
}
\details{
When representing a interaction pair, the names of both clusters involved are put together,
and kClustersSplit is the only char used to split them.
Please make sure that kClustersSplit does not appear in any name of clusters, or you can 
modify this by hand.
}
\keyword{datasets}
