% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalvariables.R
\docType{data}
\name{kGenesSplit}
\alias{kGenesSplit}
\title{Genes split character in one gene pair (global).}
\format{
An object of class \code{character} of length 1.
}
\usage{
kGenesSplit
}
\description{
This is the character used to split gene names in gene pairs, e.g EGF>EGFR.
}
\details{
When representing a gene pair, the names of both genes participated are put together,
and kGenesSplit is the character used to split them.
Please make sure that kGenesSplit does not appear in any name of genes, or you can 
modify this by hand.
}
\keyword{datasets}
