% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\docType{data}
\name{kmusthave.colnames}
\alias{kmusthave.colnames}
\title{Predefined data format in given fgenes}
\format{
An object of class \code{character} of length 4.
}
\usage{
kmusthave.colnames
}
\description{
Given data on differentially expressed genes and their belonging clusters should be in the right format.
Some columns must be maintained, which will be further used in downstream analysis.
}
\keyword{datasets}
