% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\docType{data}
\name{kpred.action.effect}
\alias{kpred.action.effect}
\title{Predefined action effect}
\format{
An object of class \code{character} of length 4.
}
\usage{
kpred.action.effect
}
\description{
This is extracted from databases depicting action effect of actions for functional usages.
}
\keyword{datasets}
