% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalvariables.R
\docType{data}
\name{kpred.exprs.change}
\alias{kpred.exprs.change}
\title{Predefined expression change selections}
\format{
An object of class \code{character} of length 4.
}
\usage{
kpred.exprs.change
}
\description{
There are 4 status that one gene pairs can have upon expression change, and this package 
defines them in these 4 short expressions.
}
\keyword{datasets}
