% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\docType{data}
\name{kpred.ext.action.effect}
\alias{kpred.ext.action.effect}
\title{Predefined extended action effect}
\format{
An object of class \code{character} of length 7.
}
\usage{
kpred.ext.action.effect
}
\description{
This depicts action effect and direction of action together.
}
\keyword{datasets}
