% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalvariables.R
\docType{data}
\name{kpred.mode}
\alias{kpred.mode}
\title{Predefined mode}
\format{
An object of class \code{character} of length 8.
}
\usage{
kpred.mode
}
\description{
This is extracted from databases depicting mode of actions for functional usages.
expression is transcriptional regulation.
}
\keyword{datasets}
