% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{ListAllClusters}
\alias{ListAllClusters}
\title{List cell clusters}
\usage{
ListAllClusters(object)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}
}
\description{
This function is to list all cell clusters.
}
