% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullView.R
\name{ListClusterGroups}
\alias{ListClusterGroups}
\title{List Cluster Groups}
\usage{
ListClusterGroups(
  object,
  use.former = FALSE,
  cluster.former = character(),
  use.latter = FALSE,
  cluster.latter = character()
)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{use.former}{It controls whether to list cluster groups relating to one or some clusters(.X) given in 
parameter \code{sel.clusters.X} of \code{link{AnalyzeInterInFullView}}.}

\item{cluster.former}{The name of clusters, which is used find all X~? interactions (X is cluters given in these parameter).}

\item{use.latter}{It controls whether to list cluster groups relating to one or some clusters(.Y) given in 
parameter \code{sel.clusters.Y} of \code{link{AnalyzeInterInFullView}}.}

\item{cluster.latter}{The name of clusters, which is used find all ?~Y interactions (Y is cluters given in these parameter).}
}
\value{
Character. Names of cluster groups.
}
\description{
This function is to help on listing all or some of cluster groups (i.e. interacting clusters).
}
\examples{
\dontrun{
 ListClusterGroups(object, use.former = TRUE, cluster.former = "T_cell")
 # the output of this will be "T_cell~Macrophage", "T_cell~B_cell" and things like that, depending on available clusters.

 ListClusterGroups(object, use.latter = TRUE, cluster.latter = "T_cell")
 # the output of this will be "Macrophage~T_cell", "B_cell~T_cell" and things like that, depending on available clusters.
}

}
