% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{ListAllGeneLocation}
\alias{ListAllGeneLocation}
\alias{ListAllGeneType}
\alias{ListAllGeneMergeType}
\alias{ListAllGeneGOTerm}
\alias{ListGeneSelectionProperty}
\alias{ListAllGeneLocation.default}
\alias{ListAllGeneLocation,InterCellDBPack-method}
\alias{ListAllGeneLocation,InterCell-method}
\alias{ListAllGeneType.default}
\alias{ListAllGeneType,InterCellDBPack-method}
\alias{ListAllGeneType,InterCell-method}
\alias{ListAllGeneMergeType.default}
\alias{ListAllGeneMergeType,InterCellDBPack-method}
\alias{ListAllGeneMergeType,InterCell-method}
\alias{ListAllGeneGOTerm.default}
\alias{ListAllGeneGOTerm,InterCellDBPack-method}
\alias{ListAllGeneGOTerm,InterCell-method}
\title{List Options for Gene Selection}
\usage{
ListAllGeneLocation(object, ...)

ListAllGeneType(object, ...)

ListAllGeneMergeType(object, ...)

ListAllGeneGOTerm(object, ...)

ListAllGeneLocation.default(object)

\S4method{ListAllGeneLocation}{InterCellDBPack}(object, ...)

\S4method{ListAllGeneLocation}{InterCell}(object, ...)

ListAllGeneType.default(object)

\S4method{ListAllGeneType}{InterCellDBPack}(object, ...)

\S4method{ListAllGeneType}{InterCell}(object, ...)

ListAllGeneMergeType.default(object)

\S4method{ListAllGeneMergeType}{InterCellDBPack}(object, ...)

\S4method{ListAllGeneMergeType}{InterCell}(object, ...)

ListAllGeneGOTerm.default(object, n.output = +Inf)

\S4method{ListAllGeneGOTerm}{InterCellDBPack}(object, ...)

\S4method{ListAllGeneGOTerm}{InterCell}(object, ...)
}
\arguments{
\item{object}{Allowed object should be in either class \code{InterCell} or class \code{InterCellDBPack}.}

\item{...}{Parameters passed to corresponding function with suffix 'default', like use \code{ListAllGeneLocation} 
and check parameters in \code{ListAllGeneLocation.default}.}

\item{n.output}{It gives the first N GO terms in the order of AlphaBet.}
}
\value{
Character. The options.
}
\description{
This group of functions are served for giving options on selecting gene subset. It ranges from 
subcellular locations, molecular functions and GO terms.
}
