% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{MergeDBSubset}
\alias{MergeDBSubset}
\title{Merge Database}
\usage{
MergeDBSubset(db.object.1, db.object.2, merge.option = "intersect")
}
\arguments{
\item{db.object.1}{A \code{InterCellDBPack} object to be merged.}

\item{db.object.2}{Another \code{InterCellDBPack} object to be merged.}

\item{merge.option}{The supported options are 'intersect' and 'union'.}
}
\value{
A \code{InterCellDBPack} object.
}
\description{
This function is to merge 2 databases to get user-desired database.
}
