% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairs.human.db}
\alias{pairs.human.db}
\title{Experiments database of interaction pairs of human genes}
\format{
Data.frame.
\itemize{
  \item inter.GeneID.*: the unique geneID for every gene in NCBI gene database.
   \item inter.GeneName.*: the authorized name for every gene in NCBI gene database.
   \item inter.Experiments.Score: score by experiments, which is the <experiments> score in STRING database.
   \item inter.Database.Score: score by curated database, which is the <database> score in STRING database.
   \item inter.Predicted.Score: score by prediction, which is the columns other than <experiments> and <database.>, 
         e.g. gene fusion, gene neighbor, etc.
   \item inter.Combined.Score: score combing all above scores.
}
}
\source{
STRING <https://string-db.org>.
}
\usage{
pairs.human.db
}
\description{
A dataset contains the all-from-STRING interaction pairs of human genes.
In this package, it has 5758309 unique pairs in total.
}
\details{
The subset of STRING data used are <taxonomy = Homo Sapiens>.
}
\keyword{datasets}
