% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairs.human.experiments.db}
\alias{pairs.human.experiments.db}
\title{Experiments database of interaction pairs of human genes}
\format{
Data.frame.
\itemize{
  \item inter.GeneID.*: the unique geneID for every gene in NCBI gene database.
   \item inter.GeneName.*: the authorized name for every gene in NCBI gene database.
   \item inter.Experiments.Score: score by experiments, which is the <experiments> score in STRING database.
   \item inter.Database.Score: score by curated database, which is the <database> score in STRING database.
   \item inter.Predicted.Score: score by prediction, which is the columns other than <experiments> and <database.>, 
         e.g. gene fusion, gene neighbor, etc.
}
}
\source{
STRING <https://string-db.org>, HGNC <https://www.genenames.org>, GeneCards <https://www.genecards.org>, etc.
}
\usage{
pairs.human.experiments.db
}
\description{
A dataset contains the experiments-based interaction pairs of human genes, 
and is extracted from STRING database. It is further matched by Ligand/Receptor/ECM databases
in this package, and as a result, it has 214368 rows of pairs in total.
}
\details{
The subset of STRING data used are <taxonomy = Homo Sapiens>, <score[experiments] > 0>.
}
\keyword{datasets}
