% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairs.mouse.knowledge.db}
\alias{pairs.mouse.knowledge.db}
\title{Knowledge database of interaction pairs of mouse genes}
\format{
See @format in \code{?pairs.mouse.experiments.db}.
}
\source{
STRING <https://string-db.org>, HGNC <https://www.genenames.org>, GeneCards <https://www.genecards.org>, etc.
}
\usage{
pairs.mouse.knowledge.db
}
\description{
A dataset contains the curated-database-based interaction pairs of mouse genes, 
and is extracted from STRING database. It is further matched by Ligand/Receptor/ECM databases
in this package, and as a result, it has 42146 rows of pairs in total.
}
\details{
The subset of STRING data used are <taxonomy = Mus Musculus>, <score[combined_score] > 699>, 
<score[experiments] = 0>, <score[database] > 699>.
}
\keyword{datasets}
