% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\docType{data}
\name{kpred.action.effect}
\alias{kpred.action.effect}
\alias{kpred.color.effect}
\alias{kpred.ext.action.effect}
\alias{kpred.color.ext.effect}
\title{Predefined Action Effect & Color Usage for It}
\format{
An object of class \code{character} of length 4.

An object of class \code{character} of length 4.

An object of class \code{character} of length 7.

An object of class \code{character} of length 7.
}
\usage{
kpred.action.effect

kpred.color.effect

kpred.ext.action.effect

kpred.color.ext.effect
}
\description{
\code{kpred.action.effect} is extracted from databases depicting action effect of actions for functional usages.

\code{kpred.color.effect} is aligned with \code{kpred.action.effect} one-to-one. In default setting,
the color used for \bold{action effect} in plotting will use colors listed in this variable.

\code{kpred.ext.action.effect} is the \bold{extended} format, which depicts action effect and direction of action together, 
and extends action effect to 7 different types.

\code{kpred.color.ext.effect} is aligned with \code{kpred.ext.action.effect} one-to-one. In default setting,
the color used for \bold{extened action effect} in plotting will use colors listed in this variable.
}
\keyword{datasets}
