% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\docType{data}
\name{kpred.action.mode}
\alias{kpred.action.mode}
\alias{kpred.color.mode}
\title{Predefined Action Mode & Color Usage for It}
\format{
An object of class \code{character} of length 8.

An object of class \code{character} of length 8.
}
\usage{
kpred.action.mode

kpred.color.mode
}
\description{
\code{kpred.action.mode} is extracted from databases depicting mode of actions for functional usages.
expression is transcriptional regulation.

\code{kpred.color.mode} is aligned with \code{kpred.action.mode} one-to-one. In default setting,
the color used for \bold{action mode} in plotting will use colors listed in this variable.
}
\keyword{datasets}
