% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{kexprs.change}
\alias{kexprs.change}
\title{Options for Expression Change}
\format{
An object of class \code{character} of length 4.
}
\usage{
kexprs.change
}
\description{
\code{kexprs.change} defines the options to be used for selecting 
gene pairs under specific expression-change status.
}
\keyword{datasets}
