% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{setRefDatabase}
\alias{setRefDatabase}
\alias{setRefDatabase,InterCell-method}
\alias{getRefDatabase}
\alias{getRefDatabase,InterCell-method}
\title{Using Reference Database}
\usage{
setRefDatabase(object, ...)

\S4method{setRefDatabase}{InterCell}(object, new.ref.database)

getRefDatabase(object, ...)

\S4method{getRefDatabase}{InterCell}(object)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{...}{Parameters passed to other methods.}

\item{new.ref.database}{A new reference database.}
}
\description{
The \code{setRefDatabase} function is to \bold{set} the reference database for \code{InterCell}.

The \code{getRefDatabase} function is to \bold{get} the reference database for \code{InterCell}.
}
\examples{
\dontrun{
  setRefDatabase(object, some.new.ref.database)
}

\dontrun{
  getRefDatabase(object)
}

}
