% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{ReplaceClusterName}
\alias{ReplaceClusterName}
\title{Replace Cluster name}
\usage{
ReplaceClusterName(object, cluster.names.current, cluster.names.replace)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{cluster.names.current}{The currently used name of cell clusters.}

\item{cluster.names.replace}{The new used name of cell clusters.}
}
\value{
A \code{InterCell} object.
}
\description{
This is function is to change the names of cell clusters.
}
\examples{
\dontrun{
  ReplaceClusterName(object, "Macrophage", "Myeloid")
}

}
