% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReverseOddEvenCols}
\alias{ReverseOddEvenCols}
\title{Permutate Reverse Order of Odds and Evens}
\usage{
ReverseOddEvenCols(ncols)
}
\arguments{
\item{ncols}{Integer. An arbitrary integer.}
}
\description{
This function generates reverse premutation of odd values and even values 
in continues values, e.g. \code{1:10}.
}
\examples{
# for 1:10 
ReverseOddEvenCols(10)
# the reuslt is c(2,1,4,3,6,5,8,7,10,9)

}
