% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{SelectDBSubset}
\alias{SelectDBSubset}
\alias{SelectDBSubset,InterCellDBPack-method}
\alias{SelectDBSubset,InterCell-method}
\alias{SelectDBSubset.default}
\title{Select Subset of Gene Pair Database}
\usage{
SelectDBSubset(object, ...)

\S4method{SelectDBSubset}{InterCellDBPack}(object, ...)

\S4method{SelectDBSubset}{InterCell}(object, ...)

SelectDBSubset.default(
  object,
  combined.score.range = c(1, 1000),
  use.exp = TRUE,
  exp.score.range = c(1, 1000),
  use.know = TRUE,
  know.score.range = c(1, 1000),
  use.pred = TRUE,
  pred.score.range = c(1, 1000),
  sel.physical = FALSE,
  sel.action.mode = "ALL",
  sel.action.effect = "ALL",
  sel.action.merge.option = "intersect",
  slim.along.with.pairs = TRUE
)
}
\arguments{
\item{object}{A \code{InterCellDBPack} or \code{InterCell} object. \code{SelectDBSubset.default} uses
\code{InterCellDBPack} as input.}

\item{...}{Parameters passed to function \code{SelectDBSubset.default}.}

\item{combined.score.range}{The combined score from all evidence sources, which works on the whole database.}

\item{use.exp}{It adds the data whose evidence is experimentally validatd.}

\item{exp.score.range}{It controls the score range when selecting subset by experimentally validatd evidence, i.e. when \code{use.exp = TRUE}. 
The score should be 2 numbers within 1~1000.}

\item{use.know}{It adds the data whose evidence is pathway curated.}

\item{know.score.range}{It controls the score range when selecting subset by pathway curated evidence, i.e. when \code{use.know = TRUE}. 
The score should be 2 numbers within 1~1000.}

\item{use.pred}{It adds the data whose evidence is predicted.}

\item{pred.score.range}{It controls the score range when selecting subset by predicted evidence, i.e. when \code{use.pred = TRUE}. 
The score should be 2 numbers within 1~1000.}

\item{sel.physical}{It selects the subset of gene pairs whose corresponding protein pairs are physical associated. This parameter is 
identical to set \code{sel.action.mode = "binding"} for now due to the database limitation. It may change in future.}

\item{sel.action.mode}{Selection by action mode. "ALL" means not use this to select subset.
Other options will be directly select gene pair in that action mode. Supported options are listed in \code{kpred.action.mode}.}

\item{sel.action.effect}{Selection by action effect. "ALL" means not use this to select subset. Other 
options will be directly select gene pair in that action effect. Supported options are listed in \code{kpred.action.effect}.}

\item{sel.action.merge.option}{Either 'intersect' or 'union'. The option for merging the result from selection on action mode and action effect.}

\item{slim.along.with.pairs}{This decides whether to select the corresponding subset of action pair database 
after selecting subset of gene pair database.}
}
\value{
A \code{InterCellDBPack} or \code{InterCell} object, which is the same as given parameter \code{object}.
}
\description{
This function is to select subset of gene pair databases, and options on evidence sources, 
confidence level, action properties (mode and effect) will be used. The result from different evidence sources
will be \bold{union} in default settings. The result from action properties could be either intersection result or union.
}
\details{
The 3 evidence channels (\code{use.exp}, \code{use.know}, \code{use.pred}) are in identical priority. 
For one gene pair, it could have scores from 3 evidence channels at the same time. In default setting, the 
result of this function will be \bold{union} of results from 3 channels. If union is not satisfied, the function 
\code{\link{MergeDBSubset}} will help.

The score is within 1~1000. Score (>=700) would be consider as high confidence, and (>=900) is even higher confident.
Score (>=400) would be consider over medium confidence.
Score (<400) would be consider low confidence, while Score (<150) would be the lowest.
}
