% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-prep.R
\name{SelectInterSubset}
\alias{SelectInterSubset}
\title{Select Some Gene Pairs from One Interaction}
\usage{
SelectInterSubset(
  object,
  direction.X.to.Y = NULL,
  sel.exprs.change = c("Xup.Yup", "Xup.Ydn", "Xdn.Yup", "Xdn.Ydn"),
  sel.some.genes.X = NULL,
  sel.some.genes.Y = NULL,
  sel.genes.option = "intersect",
  sel.gene.pairs = NULL,
  sel.action.mode = "ALL",
  sel.action.effect = "ALL",
  sel.action.merge.option = "intersect"
)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{direction.X.to.Y}{Options are 'NULL', 'TRUE', 'FALSE'. It selects subset of data based on direction.
The 'NULL' will keep 2-way interacting pairs, 'TRUE' keeps the X-to-Y pairs and 'FALSE' keeps the Y-to-X pairs.}

\item{sel.exprs.change}{Options are 'Xup.Yup', 'Xup.Ydn', 'Xdn.Yup', 'Xdn.Ydn'.
It gives the corresponding expression change of every interacting gene pair.}

\item{sel.some.genes.X}{It gives the target genes expected to be expressed by cluster X (
the same as parameter \code{cluster.x} in \code{\link{FetchInterOI}}).}

\item{sel.some.genes.Y}{It gives the target genes expected to be expressed by cluster Y (
the same as parameter \code{cluster.y} in \code{\link{FetchInterOI}}).}

\item{sel.genes.option}{Options are 'intersect' or 'union'. 'intersect' strictly restricts gene pair to 
have one gene partner in \code{sel.some.genes.X} and the other in \code{sel.some.genes.Y}. 'union' restricts 
gene pair to have at least one gene either in \code{sel.some.genes.X} or \code{sel.some.genes.Y}.}

\item{sel.gene.pairs}{Directly specify the desired gene pairs. It should be given in standard table that is generated 
by \code{\link{FormatCustomGenePairs}}. The genes in "inter.*.A" will be applied to cluster X, and "inter.*.B" will be applied to cluster Y.}

\item{sel.action.mode}{Selection by action mode. "ALL" means not use action mode to select subset.
Other options will be directly select gene pair in that action mode. Supported options are listed in \code{kpred.action.mode}.}

\item{sel.action.effect}{Selection by action effect. "ALL" means not use action effect to select subset. 
Other options will be directly select gene pair in that action effect. Supported options are listed in \code{kpred.action.effect}.}

\item{sel.action.merge.option}{Either 'intersect' or 'union'. The option for merging the result from selection on action mode and action effect.}
}
\description{
This function is to further select subset of interacting gene pairs in given interaction
}
