% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing1-analysis.R
\name{Testing.FindCommonChangedGenesInOneCluster}
\alias{Testing.FindCommonChangedGenesInOneCluster}
\title{Find common changed genes in one cluster}
\usage{
Testing.FindCommonChangedGenesInOneCluster(
  interact.pairs.acted,
  cluster.name,
  cluster.sender.or.receiver = TRUE,
  common.change.confidence = 0.8,
  cmp.extend.positive = 0.25,
  cmp.extend.negative = 0.25,
  ...
)
}
\arguments{
\item{interact.pairs.acted}{List. The return value of \code{\link{AnalyzeClustersInteracts}}.}

\item{cluster.name}{Character. The name of the cluster to be explored.}

\item{cluster.sender.or.receiver}{Character. It sets the position of the cluster in interacting clusters. 
The cluster to be explored is either sender or receiver. For gene pair A-B, the cluster is set as sender, then 
all genes in the position A of the gene pairs that the cluster participates will be collected and calculated to get the result.}

\item{common.change.confidence}{Numeric. It specifies how common the changed genes are. One gene is regarded as common changed 
gene if it are the top special genes among several interacting clusters associated with the target cluster specified in the parameter \code{cluster.name}.}

\item{cmp.extend.positive}{Numeric. It sets the collected range of special genes with positive evaluation results while evaluating by function 
\code{FindSpecialGenesInOnepairCluster}. Please refer to that function on the detailed meaning of positive evaluation result.}

\item{cmp.extend.negative}{Numeric. It sets the collected range of special genes with negative evaluation results while evaluating by function 
\code{FindSpecialGenesInOnepairCluster}. Please refer to that function on the detailed meaning of negative evaluation result.}

\item{...}{A set of parameters used in \code{FindSpecialGenesInOnepairCluster} will be passed.}
}
\value{
Character.
Names of common changed genes.
}
\description{
This function finds the common changed genes in one cluster. In other words, 
the common changed genes are collected as the intersection of special genes of one cluster in all its 
interacting clusters.
}
\details{
The common changed genes are generated from the special genes of some pairs of interacting clusters. 
One gene is regarded as common changed gene if it are the top special genes among several pairs of interacting clusters.
}
