% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing1-analysis.R
\name{Testing.PlotInteractsInMultipleClusters}
\alias{Testing.PlotInteractsInMultipleClusters}
\title{Draw Circos plot for multiple interaction pairs}
\usage{
Testing.PlotInteractsInMultipleClusters(
  interact.pairs.acted,
  fgenes.remapped.all,
  actions.ref.db,
  return.table.vals = FALSE,
  select.interacts = NULL,
  clusters.select.auto = NULL,
  direction.A.to.B = TRUE,
  if.ignore.annos = TRUE,
  sel.mode.val = NULL,
  sel.action.effect.val = NULL,
  show.legend = FALSE,
  cluster.label.cex = 0.6,
  cluster.colour = c("#fb8072", "#80b1d3", "#fdb462", "#ffffb3", "#bebada", "#b3de69",
    "#fccde5", "#8dd3c7", "#d9d9d9"),
  link.cor.colour = c("grey"),
  link.cor.border.colour = c("grey"),
  link.cor.arrow.drawn = c(TRUE, TRUE, TRUE, FALSE),
  link.cor.arrow.type = c("triangle", "ellipse", "curved", "circle"),
  link.cor.arrow.length = c(0.4, 0.3, 0.3, 0.2),
  ...
)
}
\arguments{
\item{interact.pairs.acted}{List. The return value of \code{\link{AnalyzeClustersInteracts}}, 
which is formatted data about target interaction pairs.}

\item{fgenes.remapped.all}{Data.frame. Informations about feature genes(markers), prefer using return value of 
\code{\link{DataPrep.RemapClustersMarkers}} and similar functions start with \code{DataPrep.} here.}

\item{actions.ref.db}{Data.frame. Reference pairs actions database provided in this package, the list is given as follows:
\itemize{
  \item For human: \code{actions.human.ref.db}
  \item For mouse: \code{actions.mouse.ref.db}
}}

\item{return.table.vals}{Logic. If TRUE, the table of interaction pairs will be given in return value, use \code{<-} to save the result.}

\item{select.interacts}{Character. Interacts names, e.g. \code{paste0("ClusterC", kClustersSplit, "ClusterD")}, and 
the direction will be from \code{"ClusterC"} to \code{"ClusterD"}.}

\item{clusters.select.auto}{Character. If parameter \code{select.interacts} is not specified, it gives all related clusters,
and all permutation of interacts between these clusters will be used.}

\item{direction.A.to.B}{Logic. this parameter gets 3 options: TRUE, FALSE, NULL. In InterCellDB-defined gene pairs, gene pairs have 2 genes, with 1 named "inter.GeneName.A", 
and the other named "inter.GeneName.B". As for interactions between specific cluster group like G1~G2, the gene A is expressed by the former, which is G1, and 
gene B is expressed by the latter, which is G2. So, the 3 options mean: With TRUE, all directional gene pairs 
that goes from the former cluster to the latter, like A->B or A--|B, will be extracted, but if it is set FALSE, then all directional gene pairs 
goes the opposite way, which is something like A<--B or A|--B, will be extracted. If NULL is set, then all directional gene pairs will be reserved.}

\item{if.ignore.annos}{Logic. It is passed to \code{GenerateVEinfos}. If TRUE, genes with different locations or types documented will
be treated as the same, and only one row information will be reserved.}

\item{sel.mode.val}{Character. If set NULL, it uses all values in global variables \code{InterCellDB::kpred.mode}, or
please specify detailed and accurate values in subset of \code{InterCellDB::kpred.mode}.}

\item{sel.action.effect.val}{Character. If set NULL, it uses all values in global variables \code{InterCellDB::kpred.action.effect}, or
please specify detailed and accurate values in subset of \code{InterCellDB::kpred.action.effect}.}

\item{show.legend}{Logic. Show legend(TRUE) or not(FALSE).}

\item{cluster.label.cex}{Numeric. The size of letters when labeling cluster names.}

\item{cluster.colour}{Character. Predefined some colours in harmony, generated by colorbrew2, 
see \url{http://colorbrewer2.org} for details.}

\item{link.cor.colour}{Character. Colour of links, whose length can be 1 or the same as another parameter \code{pred.mode}.}

\item{link.cor.border.colour}{Character. Colour of link borders, whose value length can be 1 or the same as another parameter \code{pred.mode}.}

\item{link.cor.arrow.drawn}{Logic. It decides if the arrows are drawn upon the links. If TRUE, draw the arrows, otherwise not.}

\item{link.cor.arrow.type}{Character. Arrow type of links, whose length must be same as another parameter \code{pred.action.effect}.}

\item{link.cor.arrow.length}{Numeric. Arrow length of links, whose length must be same as another parameter \code{pred.action.effect}.}
}
\description{
This function draws circos plot for interaction pairs of several target clusters, and returns
the detailed result tables about these.
}
\details{
This function draws circos plot to show multiple clusters invovled interaction pairs. The final circos plot
contains 2 tracks and 1 inner links gragh. (If you are not familiar with those terms, please see \url{} for details.)
\itemize{
  \item {base track: } {It represents all clusters that are drawn in different colours. 
        The colours are specified by parameter \code{cluster.colour}}
  \item {expression level track: } {It gives all genes' expression levels by values of "LogFC".
        The up-regulated ones are marked in yellowish, and the down-regulated ones are marked in green.}
  \item {links: } {It draws links pair-by-pair, and shows mode and action type of every interaction pair.}
}
}
