% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{setTgActionComp}
\alias{setTgActionComp}
\alias{setTgActionComp,InterCell-method}
\alias{getTgActionComp}
\alias{getTgActionComp,InterCell-method}
\title{Using TargetView Action Composition}
\usage{
setTgActionComp(object, ...)

\S4method{setTgActionComp}{InterCell}(object, new.action.comp)

getTgActionComp(object, ...)

\S4method{getTgActionComp}{InterCell}(object)
}
\arguments{
\item{new.action.comp}{A new result of intercellular analysis on composition of action mode and action effect.}
}
\description{
The \code{setTgActionComp} function is to \bold{set} the result of intercellular analysis
on composition of action mode and action effect for one interaction between 2 cell clusters.

The \code{getTgActionComp} function is to \bold{get} the result of intercellular analysis
on composition of action mode and action effect for one interaction between 2 cell clusters.
}
\examples{
\dontrun{
  setTgActionComp(object, some.new.action.comp)
}

\dontrun{
  getTgActionComp(object)
}
}
