% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{setTgActionPairs}
\alias{setTgActionPairs}
\alias{getTgActionPairs}
\alias{getTgActionPairs,InterCell-method}
\title{Using TargetView Action Pairs}
\usage{
setTgActionPairs(object, ...)

getTgActionPairs(object, ...)

\S4method{getTgActionPairs}{InterCell}(object)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{...}{Parameters passed to other methods.}
}
\description{
The \code{setTgActionPairs} function is to \bold{set} the result of intercellular analysis 
on action properties for one interaction between 2 cell clusters.

The \code{getTgActionPairs} function is to \bold{get} the result of intercellular analysis 
on action properties for one interaction between 2 cell clusters.
}
\examples{
\dontrun{
  getTgActionPairs(object)
}
}
