% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{setTgItInfo}
\alias{setTgItInfo}
\alias{setTgItInfo,InterCell-method}
\alias{getTgItInfo}
\alias{getTgItInfo,InterCell-method}
\title{Using Target Plain Information}
\usage{
setTgItInfo(object, ...)

\S4method{setTgItInfo}{InterCell}(object, new.itinfo)

getTgItInfo(object, ...)

\S4method{getTgItInfo}{InterCell}(object)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{...}{Parameters passed to other methods.}

\item{new.itinfo}{A new plain information on gene pairs, which are directly fetched from full view analysis.}
}
\description{
The \code{setTgItInfo} function is to \bold{set} plain information about gene pairs,
which are directly fetched from full view analysis.

The \code{getTgItInfo} function is to \bold{get} plain information about gene pairs,
which are directly fetched from full view analysis.
}
\examples{
\dontrun{
  setTgItInfo(object, some.new.itinfo)
}

\dontrun{
  getTgItInfo(object)
}
}
