% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{setTgSpGenes}
\alias{setTgSpGenes}
\alias{setTgSpGenes,InterCell-method}
\alias{getTgSpGenes}
\alias{getTgSpGenes,InterCell-method}
\title{Using TargetView Pair Specificity}
\usage{
setTgSpGenes(object, ...)

\S4method{setTgSpGenes}{InterCell}(object, new.spgenes)

getTgSpGenes(object, ...)

\S4method{getTgSpGenes}{InterCell}(object)
}
\arguments{
\item{new.spgenes}{A new result of intercellular analysis on gene pair specificity.}
}
\description{
The \code{setTgSpGenes} function is to \bold{set} the result of intercellular analysis
on gene pair specificity for one interaction between 2 cell clusters.

The \code{getTgSpGenes} function is to \bold{get} the result of intercellular analysis
on gene pair specificity for one interaction between 2 cell clusters.
}
\examples{
\dontrun{
  setTgSpGenes(object, some.new.spgenes)
}

\dontrun{
  getTgSpGenes(object)
}
}
