% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{setTgVEInfo}
\alias{setTgVEInfo}
\alias{setTgVEInfo,InterCell-method}
\alias{getTgVEInfo}
\alias{getTgVEInfo,InterCell-method}
\title{Using TargetView VEinfo}
\usage{
setTgVEInfo(object, ...)

\S4method{setTgVEInfo}{InterCell}(object, new.veinfo)

getTgVEInfo(object, ...)

\S4method{getTgVEInfo}{InterCell}(object)
}
\arguments{
\item{new.veinfo}{A new result of intercellular analysis on detailed gene pairs.}
}
\description{
The \code{setTgVEInfo} function is to \bold{set} the result of intercellular analysis
on detailed gene pairs (forming the interaction network) for one interaction between 2 cell clusters.

The \code{getTgVEInfo} function is to \bold{get} the result of intercellular analysis
on detailed gene pairs (forming the interaction network) for one interaction between 2 cell clusters.
}
\examples{
\dontrun{
  setTgVEInfo(object, some.new.veinfo)
}

\dontrun{
  getTgVEInfo(object)
}
}
