% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{TgView.formula.onLogFC.default}
\alias{TgView.formula.onLogFC.default}
\title{Formula on Power for One Gene Pair}
\usage{
TgView.formula.onLogFC.default(data.f, data.b)
}
\arguments{
\item{data.f}{vector. The LogFC values for one list of gene partners. LogFC values are 
specified when creating \code{\link{InterCell}} object.}

\item{data.b}{vector. The LogFC values for another list of gene partners, which is 
one-by-one matched to those in parameter \code{data.f}.}
}
\description{
This is the default formula on calculating power for every gene pair, and the 
sum value of LogFC for gene partners will be regarded as power.
}
