% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{TgView.formula.onPVal.default}
\alias{TgView.formula.onPVal.default}
\title{Formula on Confidence for One Gene Pair}
\usage{
TgView.formula.onPVal.default(data.f, data.b, pval.log.max = 1000)
}
\arguments{
\item{data.f}{vector. The PVal values for one list of gene partners. PVal values are 
specified when creating \code{\link{InterCell}} object.}

\item{data.b}{vector. The PVal values for another list of gene partners, which is 
one-by-one matched to those in parameter \code{data.f}.}
}
\description{
This is the default formula on calculating confidence for every gene pair, and the 
multiplying value of PVal for gene partners will be regarded as confidence.
}
