% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-extension-usage.R
\name{Tool.AddUserRestrictDB}
\alias{Tool.AddUserRestrictDB}
\title{Add user-defined dataset}
\usage{
Tool.AddUserRestrictDB(user.def.db, genes.ref.db, warning.given = "genes")
}
\arguments{
\item{user.def.db}{Data.frame. It should be a data.frame with at least 2 columns, whose 1st column is gene names, and the 2nd is 
definitions in users' personal will, and colums from the 3rd will be inherited intactly.}

\item{genes.ref.db}{List. It is taxonomy specific, which contains 3 parts:
1. gene.ncbi.db            NCBI-gene database for specific species
2. gene.synonyms.db        GeneID-GeneName-Synonym pairs extracted from (1)
3. gene.dup.synonyms.db  Some genes may have same synonyms, and this can cause mistakes 
when match target synonym back to its authorized genename.}

\item{warning.given}{Character. It applies extra limitation on the types(molecular functions) of A in gene pairs formatted as A-B.}
}
\value{
Data.frame.
| <col-1> <col-2> <col-3> ...|

| GeneID Gene.name user.type ...|

| <data> <data> <data> ...|
}
\description{
This function gives a standlized procedure that package users can easily add its 
definitions about genes and their functions or types or locations etc.
}
