% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-extension-usage.R
\name{Tool.FindGenesFromGO}
\alias{Tool.FindGenesFromGO}
\title{Find genes annotated in specific GO terms}
\usage{
Tool.FindGenesFromGO(go.todolist, genes.ref.db, go.ref.db)
}
\arguments{
\item{go.todolist}{Character. Several GO_terms or GO_IDs or mixed, which will 
be used to get subsets of feature genes.}

\item{genes.ref.db}{List. It is taxonomy specific, which contains 3 parts:
1. gene.ncbi.db            NCBI-gene database for specific species
2. gene.synonyms.db        GeneID-GeneName-Synonym pairs extracted from (1)
3. gene.dup.synonyms.db  Some genes may have same synonyms, and this can cause mistakes 
when match target synonym back to its authorized genename.}

\item{go.ref.db}{Data.frame. Reference GO database provided in this package, the list is given as follows:
\itemize{
  \item For human: \code{go.human.ref.db}
  \item For mouse: \code{go.mouse.ref.db}
}}
}
\value{
Character. A gene list of given GO IDs or terms.
}
\description{
This function uses GO terms or GO IDs to get a specific list of genes.
}
