% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-extension-usage.R
\name{Tool.formula.onPValAdj.default}
\alias{Tool.formula.onPValAdj.default}
\title{Formula on calculating p_val_adj}
\usage{
Tool.formula.onPValAdj.default(data.f, data.b)
}
\arguments{
\item{data.f}{vector. The operated data.}

\item{data.b}{vector. The operated data with its length the same as the 1st parameter.}
}
\description{
This is the default formula on calculating p_val_adj values, and is 
belong to \code{GetResult.PlotOnepairClusters.GeneCrosstalk()}.
}
\details{
This is the default formula for calculation. Users can define their own formula,
as well as define their own evalution parameter besides p_val_adj(by changing the 
param - \code{colnames.to.cmp} of \code{GetResult.PlotOnepairClusters.GeneCrosstalk()}).
}
