% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Tool.GenPermutation}
\alias{Tool.GenPermutation}
\title{Generate Permutations of Count Matrix}
\usage{
Tool.GenPermutation(
  object,
  count.matrix,
  cells.meta = NULL,
  genes.name = NULL,
  perm.times = 1000
)
}
\arguments{
\item{object}{A \code{InterCell} object.}

\item{count.matrix}{Normalized count matrix, with columns as cell barcodes and rows as gene names.}

\item{cells.meta}{cell group labels, which are used to group cells in several groups rather all different barcodes.}

\item{genes.name}{Readable gene names, which should be the same set as genes given in \code{DEG.table}.
To be noted, if genes are remapped when created \code{InterCell} object, it will be automatically applied to this.}

\item{perm.times}{Permutation times. Default is set to 1000.}
}
\description{
This function uses cell label permutation to generate one list of normalized count matrix.
}
