% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-extension-usage.R
\name{Tool.GenStdGenePairs.from.VEinfos}
\alias{Tool.GenStdGenePairs.from.VEinfos}
\title{generate gene pairs in standard format from VEinfos}
\usage{
Tool.GenStdGenePairs.from.VEinfos(VEinfos)
}
\arguments{
\item{VEinfos}{List. It contains informations about vertices and edges, and is exactly return value of
\code{GenerateVEinfos()} or \code{TrimVEinfos()}.}
}
\description{
This function generates gene pairs in standard format(in data frame), 
and gets these pairs easier to be compared with others.
}
\details{
The standard format in this package is that gene pairs are maintained in data.frame, and the 2 genes 
participated in each gene pair are recorded in columns named "inter.GeneName.A" and "inter.GeneName.B".
}
