% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-output.R
\name{Tool.ShowGraph}
\alias{Tool.ShowGraph}
\title{Tool for showing gragh}
\usage{
Tool.ShowGraph(formatted.result, ...)
}
\arguments{
\item{formatted.result}{List. It must have one item named "plot".}

\item{...}{other plot functions that directly passed using \code{+} in definition of \pkg{ggplot2}.
It will make slight changes on the final plotting result.}
}
\description{
This function is a simple interface to show the graph in results that are 
from functions \code{GetResult.*}.
}
