% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-common.R
\name{Tool.SplitToGenDataFrame}
\alias{Tool.SplitToGenDataFrame}
\title{split character to be data.frame}
\usage{
Tool.SplitToGenDataFrame(to.splits.string, to.split.by, res.colnames)
}
\arguments{
\item{to.splits.string}{[TODO]}

\item{to.split.by}{[TODO]}

\item{res.colnames}{[TODO]}
}
\description{
This function generates n-column data frame by spliting character by some specific letters or phrases.
}
