% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-output.R
\name{Tool.WriteTables}
\alias{Tool.WriteTables}
\title{Tool for write tables in csv}
\usage{
Tool.WriteTables(formatted.result, dir.path = ".", ...)
}
\arguments{
\item{formatted.result}{List. It must have one item named "table".}

\item{dir.path}{Character. The written directory path in operating system.}

\item{...}{Other params that will be directly passed to write.csv, use \code{?write.csv} for 
further help.}
}
\description{
This function is a simple interface to write the tables in results that are 
from functions \code{GetResult.*}
}
