% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-in-targets.R
\name{TrimVEinfos}
\alias{TrimVEinfos}
\title{Trim the vertices and edges infos}
\usage{
TrimVEinfos(
  VEinfos,
  sel.some.gene.pairs.df = NULL,
  sel.some.gene.pairs.colnames = c("inter.GeneName.A", "inter.GeneName.B"),
  sel.exprs.change = c("Xup.Yup", "Xup.Ydn", "Xdn.Yup", "Xdn.Ydn"),
  direction.A.to.B = NULL,
  sel.mode.val = NULL,
  sel.action.effect.val = NULL,
  sel.mode.action.merge.option = "intersect"
)
}
\arguments{
\item{VEinfos}{List. It contains informations about vertices and edges, and is exactly return value of
\code{GenerateVEinfos()} or \code{TrimVEinfos()}.}

\item{sel.some.gene.pairs.df}{Data.frame. It is at-least-2-column data.frame, which records gene pairs with each column settling 
one of the participated genes. The 2 required column names need to be specified by another parameter \code{sel.some.gene.pairs.colnames}.}

\item{sel.some.gene.pairs.colnames}{Character of length 2. It strictly specifies the column names that records the genes of given gene pairs, 
and it also implies the direction goes from the first column to the second (AtoB). So, make sure putting genes in their proper positions.}

\item{sel.exprs.change}{Character. It selects the expression change status that gene pairs can be. It has total 4 options:
"Xup.Yup", "Xup.Ydn", "Xdn.Yup", "Xdn.Ydn", which are defined in global variables \code{kpred.exprs.change}.}

\item{direction.A.to.B}{Logic. this parameter gets 3 options: TRUE, FALSE, NULL. In InterCellDB-defined gene pairs, gene pairs have 2 genes, with 1 named "inter.GeneName.A", 
and the other named "inter.GeneName.B". As for interactions between specific cluster group like G1~G2, the gene A is expressed by the former, which is G1, and 
gene B is expressed by the latter, which is G2. So, the 3 options mean: With TRUE, all directional gene pairs 
that goes from the former cluster to the latter, like A->B or A--|B, will be extracted, but if it is set FALSE, then all directional gene pairs 
goes the opposite way, which is something like A<--B or A|--B, will be extracted. If NULL is set, then all directional gene pairs will be reserved.}

\item{sel.mode.val}{Character. If set NULL, it uses all values in global variables \code{InterCellDB::kpred.mode}, or
please specify detailed and accurate values in subset of \code{InterCellDB::kpred.mode}.}

\item{sel.action.effect.val}{Character. If set NULL, it uses all values in global variables \code{InterCellDB::kpred.action.effect}, or
please specify detailed and accurate values in subset of \code{InterCellDB::kpred.action.effect}.}

\item{sel.mode.action.merge.option}{Character. Its allowed values are "intersect" and "union". It defines the way that merges the result of subset selected by mode 
and subset selected by action.effect. The default way is to intersect the subsets, and the other option is union.}
}
\description{
This function uses mode and action.effect as the restrictions when users select part of the result.
}
\details{
The whole list of mode or action.effect is defined as global variable that is given within the package.
kpred.mode defines 9 modes, while kpred.action.effect defines 4 action effects.
As the action database given now is not completed well, it is recommended to select upon
action.effect, while leaving all different modes preserved.
}
