% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Uniprot.key.map.list}
\alias{Uniprot.key.map.list}
\title{Merged Uniprot Keywords}
\format{
Data.frame.
\itemize{
   \item {\code{orig.uniprot.keywords}: } {[TODO]}
   \item {\code{merged.molecular.function}: } {[TODO]
         }
}
}
\source{
<https://www.uniprot.org>, and download the Keywords-[molecular function] of genes.
}
\usage{
Uniprot.key.map.list
}
\description{
A dataset contains the mapping of original Uniprot Keywords and remapped keys
}
\keyword{datasets}
