% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-part1.R
\name{AnalyzeInterInAction}
\alias{AnalyzeInterInAction}
\title{Collect gene pairs by expression change and action effects}
\usage{
AnalyzeInterInAction(
  object,
  limits.exprs.change = kexprs.change,
  limits.action.mode = kpred.mode,
  limits.action.effect = kpred.action.effect
)
}
\arguments{
\item{object}{[TODO]}

\item{limits.exprs.change}{It specifies the range of expression change status that needs to be collected.}

\item{limits.action.mode}{[TODO]}

\item{limits.action.effect}{It specifies the range of action effects that need to be collected.}
}
\value{
[TODO]
}
\description{
This function gets to collect gene pairs by their expression change (total 4 category) and 
action effects (total 4 category).
}
\details{
This function is to collect hierachical information inside the raw data, which is collecting gene pairs 
and grouping them by expression change status and action effects.
}
