% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-part3.R
\name{AnalyzeInterSpecificity}
\alias{AnalyzeInterSpecificity}
\title{Analyze Gene Pairs by Specificity}
\usage{
AnalyzeInterSpecificity(
  object,
  to.cmp.cluster.groups = character(),
  extended.search = FALSE,
  uq.cnt.options = c(1:100)
)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{to.cmp.cluster.groups}{It gives the interactions (cluster groups) to be compared. 
\code{\link{ListClusterGroups}} helps to give the names of interactions.}

\item{extended.search}{Decide whether to add more gene pairs (those from compared cluster groups) into analysis.}

\item{uq.cnt.options}{Integer. It defines the allowed co-occurence count of one gene pair, and count goes larger impling 
lower specificity. The recommended value is \code{1:(length(to.cmp.cluster.groups) + 1)}.}
}
\value{
A \code{InterCell} object.
}
\description{
This function is used to evaluate specificity for gene pairs from one interaction. In other word, 
it explores every gene pair with all its occuring interactions. The one interaction is controlled by 
\code{\link{FetchInterOI}}.
}
