% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-in-targets.R
\name{CollectHierachyOnepairClusters}
\alias{CollectHierachyOnepairClusters}
\title{TODO}
\usage{
CollectHierachyOnepairClusters(onepair.gmoc)
}
\arguments{
\item{onepair.gmoc}{List. Return value of GenerateMapDetailOnepairClusters().}
}
\value{
TODO
}
\description{
TODO
}
\details{
This function is to collect hierachical information inside the raw data.
}
