% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{CreateInterCellObject}
\alias{CreateInterCellObject}
\title{Analyze interaction network in full view}
\usage{
CreateInterCellObject(
  DEG.table,
  species,
  cluster.split = "~",
  gene.pair.split = "-~-",
  remap.genes = FALSE
)
}
\arguments{
\item{DEG.table}{A table on differentially expressed genes and their belonging cell clusters.}

\item{species}{\pkg{InterCellDB} supports species in either 'human' or 'mouse'.
Other species are not supported to be used for analysis.}

\item{cluster.split}{The letters used to split 2 cell clusters in one interaction. It can also be modified later by 
using \code{setClusterSplit} if not decided yet.}

\item{gene.pair.split}{The letters used to split 2 gene partners in one gene pair. It can also be modified later by 
using \code{setGenePairSplit} if not decided yet.}

\item{remap.genes}{This decides whether to use \pkg{InterCellDB} integrated gene database to
standardize all genes to get more perfectly matched with protein/gene pairs database. Default is \code{FALSE}.}
}
\value{
A \code{InterCell} object.
}
\description{
This function analyzes count and power of interaction pairs among all given clusters.
}
\details{
The parameter \code{DEG.table} is recommended to be generated by \pkg{Seurat}. Other packages are also applicable, if
they work with scRNA-seq, cell clustering and calculation on cluster-specific differentially expressed genes. The input 
format of \code{DEG.table} should be one data.frame with 4 required columns that are named 'cluster', 'gene', 'LogFC', 'PVal'.
\itemize{
  \item{cluster}: the cell cluster.
  \item{gene}: differentially expressed genes, which are grouped by their belonging clusters.
  \item{LogFC}: the fold change of genes.
  \item{PVal}: the P value for gene being calculated as differentially expressed gene.
}

To represent one interaction, like interaction between 'Myeloid cell' and 'T cell', 
it will be looked like 'Myeloid cell~T cell' if \code{cluster.split = "~"}.
To represent oen gene pair, like gene pair of IL6 and IL6R,
it will be looked like 'IL6-~-IL6R' if \code{gene.pair.split = "-~-"}.

To avoid program failure, the letters appearing in gene names are not recommended for \code{gene.pair.split},
and the letters appearing in cluster names are not recommended for \code{cluster.split}. 
The program will test for those situation but users should keep this in mind.
}
