% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preparation.R
\name{DataPrep.RemapClustersMarkers}
\alias{DataPrep.RemapClustersMarkers}
\title{Remap gene symbols to their authorized genenames}
\usage{
DataPrep.RemapClustersMarkers(
  markers.all.from.Seurat,
  genes.ref.db,
  warning.given = "markers",
  if.used.inside = FALSE
)
}
\arguments{
\item{markers.all.from.Seurat}{Data.frame. Feature genes that are from return value of \code{Seurat::FindAllMarkers()}.}

\item{genes.ref.db}{List. It is taxonomy specific, which contains 3 parts:
1. gene.ncbi.db            NCBI-gene database for specific species
2. gene.synonyms.db        GeneID-GeneName-Synonym pairs extracted from (1)
3. gene.dup.synonyms.db  Some genes may have same synonyms, and this can cause mistakes 
when match target synonym back to its authorized genename.}

\item{warning.given}{Character. It substitutes a small part of the warning sentences. In 
most cases, you just leave it out.}
}
\description{
Use species-specific reference database to remap genes to their authorized genenames
}
