% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{FetchGeneOI}
\alias{FetchGeneOI}
\alias{FetchGeneOI,InterCellDBPack-method}
\alias{FetchGeneOI,InterCell-method}
\alias{FetchGeneOI.default}
\title{Fetch Genes of Interest}
\usage{
FetchGeneOI(object, ...)

\S4method{FetchGeneOI}{InterCellDBPack}(object, ...)

\S4method{FetchGeneOI}{InterCell}(object, ...)

FetchGeneOI.default(
  object,
  sel.location = NULL,
  sel.location.score = c(1:5),
  sel.type = NULL,
  sel.merge.type = NULL,
  sel.go.terms = NULL,
  go.use.relative = TRUE,
  go.relative.option = "offspring"
)
}
\arguments{
\item{object}{A \code{InterCell} object or \code{InterCellDBPack} object. \code{FetchGeneOI.default}
gets \code{InterCellDBPack} as input.}

\item{...}{Parameters passed to function \code{FetchGeneOI.default}.}

\item{sel.location}{Use subcellular location of gene product to select gene, and options are listed 
in \code{\link{ListAllGeneLocation}}.}

\item{sel.location.score}{The score of corresponding subcellular location, range from 1 to 5. Consider \code{score = c(4, 5)} 
as high and common usage.}

\item{sel.type}{Use molecular function of gene product to select gene, and options are listed in 
\code{\link{ListAllGeneType}}.}

\item{sel.merge.type}{The merged types. Comparing to \code{sel.type}, it has less options, which are given 
in \code{\link{ListAllGeneMergeType}}. See details for help.}

\item{sel.go.terms}{Use GO terms to select gene, and supported options are listed in \code{\link{ListAllGeneGOTerm}}.}

\item{go.use.relative}{Decide if the go terms in the related GO term tree are used. See details for help.}

\item{go.relative.option}{Decide which relation to the selected GO term is used. Options are 
'ancestor', 'parents', 'offspring', 'children'. See details for help.}
}
\value{
Character. The selected genes.
}
\description{
This function is to fetch genes of interest by selecting on subcellular locations, 
molecular functions and GO terms.
}
\details{
The parameter \code{sel.merge.type} is the summary for \code{sel.type}. The options in \code{sel.type} are originally 
generated directly from Uniprot, which comprises over 100 types. For the convenience of usage, we summarize those types and gather
them to 16 merged types, which comprise the common used types: 'Receptor', 'Cytokine', 'Growth Factor', etc.

GO terms have 3 basic words: 'cellular_component', 'molecular_function', 'biological_process', and all other GO terms are the offspring of 
one of them. As a result, GO terms form 3 family tree. Use parameter \code{go.use.relative}, it can extend one given GO term to all its 
related GO terms. There are 4 pre-defined options for selecting specific relative group, which are 'ancestor', 'parents', 'offspring', 'children'.
The 'parents' and 'children' are selecting the most close GO terms. For example, 'GO:0006955-immune response' is the nearest level above 
'GO:0002250-adaptive immune response', and in turn, 'adaptive immune response' is the children of 'immune response'. 
The 'ancestor' and 'offspring' goes further than 'parents' and 'children'. The 'ancestor' iteratively searches for the 'parents'. 
Conversely, the 'offspring' iteratively searches for the 'children'. For example, 'immune response' is the parents of 'adaptive immune response', and 
'adaptive immune response' is the parents of 'adaptive immune effector response'. Then, 'immune response' is the 'ancestor' of 'adaptive immune effector response'.
The 'offspring' goes the same way by propagating 'children'.
}
