% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-find-specialgenes.R
\name{FindSpecialGenesInOnepairCluster}
\alias{FindSpecialGenesInOnepairCluster}
\title{Find special genes in one pair of interacting clusters}
\usage{
FindSpecialGenesInOnepairCluster(
  interact.pairs.acted,
  target.gene.pairs.df,
  involved.clusters.pair,
  to.cmp.clusters.pairs = character(),
  to.cmp.clusters.pairs.sel.strategy = "inter-cross",
  merge.cnt.quantile = 0.5,
  merge.cnt.decreasing = TRUE,
  merge.cnt.topn.shown = 20,
  plot.fill = c("red", "green"),
  sep.inside.gene.pair = "->"
)
}
\arguments{
\item{interact.pairs.acted}{List. The return value of \code{\link{AnalyzeClustersInteracts}}.}

\item{target.gene.pairs.df}{[TODO]}

\item{involved.clusters.pair}{[TODO]}

\item{to.cmp.clusters.pairs}{[TODO]}

\item{to.cmp.clusters.pairs.sel.strategy}{[TODO]}

\item{merge.cnt.quantile}{[TODO]}

\item{merge.cnt.decreasing}{[TODO]}

\item{merge.cnt.topn.shown}{[TODO]}

\item{plot.fill}{[TODO]}

\item{sep.inside.gene.pair}{[TODO]}
}
\value{
List. [TODO]
}
\description{
This function is used to find special genes in one pair of interacting clusters. Genes are special
if it passes some limitations when comparing to other pairs of interacting clusters.
}
\details{
[TODO]
}
