% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed-view-part2.R
\name{FindSpecialGenesInOnepairCluster}
\alias{FindSpecialGenesInOnepairCluster}
\title{Find special genes in one pair of interacting clusters}
\usage{
FindSpecialGenesInOnepairCluster(
  VEinfos,
  interact.pairs.acted,
  to.cmp.clusters.pairs = character(),
  to.cmp.clusters.pairs.sel.strategy = "inter-cross",
  uq.cnt.range = c(1:2),
  formula.to.use.onLogFC = Tool.formula.onLogFC.default
)
}
\arguments{
\item{VEinfos}{List. It contains informations about vertices and edges, and is exactly return value of
\code{GenerateVEinfos()} or \code{TrimVEinfos()}.}

\item{interact.pairs.acted}{List. The return value of \code{\link{AnalyzeClustersInteracts}}.}

\item{to.cmp.clusters.pairs}{[TODO]}

\item{to.cmp.clusters.pairs.sel.strategy}{[TODO]}

\item{uq.cnt.range}{[TODO]}

\item{formula.to.use.onLogFC}{[TODO]}
}
\value{
List. [TODO]
}
\description{
This function is used to find special genes in one pair of interacting clusters. Genes are special
if it passes some limitations when comparing to other pairs of interacting clusters.
}
\details{
[TODO]
}
