% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-find-specialgenes.R
\name{FindSpecialGenesInOnepairCluster}
\alias{FindSpecialGenesInOnepairCluster}
\title{Find special genes in one pair of interacting clusters}
\usage{
FindSpecialGenesInOnepairCluster(
  interact.pairs.acted,
  clusters.onepair.select,
  merge.confidence.on.diff = 0.8,
  merge.confidence.on.shared = 0.8,
  twist.fold.change.mul = 1,
  top.ignored.genes = character(),
  top.n.score.positive = 10,
  top.n.score.negative = 10,
  option.calc.score = 0,
  ...
)
}
\arguments{
\item{interact.pairs.acted}{A list. The return value of \code{\link{AnalyzeClustersInteracts}}.}

\item{clusters.onepair.select}{TODO}

\item{merge.confidence.on.diff}{TODO}

\item{merge.confidence.on.shared}{TODO}

\item{twist.fold.change.mul}{TODO}

\item{top.ignored.genes}{TODO}

\item{option.calc.score}{Integer. Defining the method use in calculating score. The default settings is: 
use \code{sum(all)}. In other cases, If it is 1, use \code{sum(abs(all))}.}

\item{...}{TODO}

\item{top.n.cnt}{TODO}

\item{top.n.score}{TODO}
}
\value{
TODO [TODO] change the import to be kind of importFrom
}
\description{
TODO
}
\details{
TODO [NOTE] as this function is based on interact.pairs.acted, if limits have been
put upon the clusters in x-axis or y-axis, the compared interacting pairs will be limited 
corresponding to the limits put upon clusters.
}
