% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCellDBclass.R
\name{FormatCustomGenePairs}
\alias{FormatCustomGenePairs}
\title{Format User Given Gene Pairs}
\usage{
FormatCustomGenePairs(object, gene.pairs.table, extend.reverse = FALSE)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{gene.pairs.table}{2-column table, and each column records one list of genes.}

\item{extend.reverse}{Decide whether to extend pairs with the reverse pairs. Default is \code{FALSE}.}
}
\value{
A list with \code{$result} storing the formatted gene pairs.
}
\description{
This function makes 2-column table (each row represents one gene pair) to be 
formatted to be used in this package.
}
\details{
The formatting process will remap the genes in 1st column from input to 'inter.*.A' columns in result.
The result of 2nd column from input will be put in 'inter.*.B' columns.

Explanation on \code{extend.reverse}:
For example, given pair C3~C3ar1, if set \code{extend.reverse = TRUE}, then 
both C3~C3ar1 and C3ar1~C3 will be generated in result.
}
