% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-in-targets.R
\name{GenerateMapDetailOnepairClusters}
\alias{GenerateMapDetailOnepairClusters}
\title{Generate interaction pairs with their actions}
\usage{
GenerateMapDetailOnepairClusters(clusters.onepair.select, actions.ref.db)
}
\arguments{
\item{clusters.onepair.select}{A list. Return value of \code{\link{ExtractTargetOnepairClusters}}.}

\item{actions.ref.db}{Data.frame. Reference pairs actions database provided in this package, the list is given as follows:
\itemize{
  \item For human: \code{actions.human.ref.db}
  \item For mouse: \code{actions.mouse.ref.db}
}}
}
\value{
A list.
\itemize{
  \item clusters.name: names of clusters involved. Its length is of 2.
  \item anno.infos: a list of lists. The sublists are 
        \itemize{
          \item location.A: it records locations of A in gene pairs formatted as A-B
          \item location.B: it records locations of B in gene pairs formatted as A-B
          \item type.A: it records molecular functions of A in gene pairs formatted as A-B
          \item type.B: it records molecular functions of B in gene pairs formatted as A-B
        }
  \item actions.detailed: A list of detailed information about interaction pairs whose actions are recorded in actions.ref.db.
}
}
\description{
This function uses actions.ref.db to distinguish as well as collect pairs whose direction of interact actions
are known in some degree and are recorded, and pairs with no detailed informations.
}
