% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-prep.R
\name{GenerateVEinfos}
\alias{GenerateVEinfos}
\title{Generate Standard Data for One Interaction}
\usage{
GenerateVEinfos(
  onepair.gmoc,
  fgenes.remapped.all,
  direction.X.to.Y = NULL,
  if.ignore.location = FALSE
)
}
\arguments{
\item{onepair.gmoc}{List. Return value of \code{\link{GenerateMapDetailOnepairClusters}}.}

\item{fgenes.remapped.all}{Data.frame. Informations about feature genes(markers), prefer using return value of 
\code{\link{DataPrep.RemapClustersMarkers}} and similar functions start with \code{DataPrep.} here.}

\item{direction.X.to.Y}{Options are 'NULL', 'TRUE', 'FALSE'. It selects subset of data based on direction.
The 'NULL' will keep 2-way interacting pairs, 'TRUE' keeps the X-to-Y pairs and 'FALSE' keeps the Y-to-X pairs.}

\item{if.ignore.location}{If set TRUE, genes with different locations or types documented will
be treated as the same, and only one row information will be reserved.}
}
\value{
A list.
\itemize{
  \item {\code{cluster.name.A}&\code{cluster.name.B}:} {cluster names involved.}
  \item edges.infos: data.frame that records the edges(the interaction pairs).
  \item vertices.infos: data.frame that records the vertices(the genes).
  \item vertices.apx.type.A: data.frame that records the types(molecular functions) of A in gene pairs formatted as A-B.
  \item vertices.apx.type.B: data.frame that records the types(molecular functions) of B in gene pairs formatted as A-B.
}
}
\description{
This function uses detailed informations about one gene pair(return value of 
\code{GenerateMapDetailOnepairClusters()}), to generate standard data with genes and 
gene pairs stored seperately.
}
\details{
This function uses actions that are recorded in STRING action database, but only a small part of 
actions are thoroughly difined in the database.
This function is used to generate formatted data structure (with vertices and edges).

In vertices, all gene informations are well recorded, and every gene is given one unique ID.

In edges, it uses unique vertices IDs to contruct the linkes, and records mode and action.effect for every link.
}
