% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed-view-in-targets.R
\name{GenerateVEinfos}
\alias{GenerateVEinfos}
\title{Generate data about vertices and edges}
\usage{
GenerateVEinfos(
  onepair.gmoc,
  fgenes.remapped.all,
  is.directional = TRUE,
  sel.mode.val = NULL,
  sel.action.effect.val = NULL
)
}
\arguments{
\item{onepair.gmoc}{List. Return value of \code{\link{GenerateMapDetailOnepairClusters}}.}

\item{fgenes.remapped.all}{Data.frame. Informations about feature genes(markers), prefer using return value of 
\code{\link{DataPrep.RemapClustersMarkers}} and similar functions start with \code{DataPrep.} here.}

\item{is.directional}{Logic. If TRUE, it only generates VEinfos from \code{onepair.gmoc$clusters.name[1]} to \code{onepair.gmoc$clusters.name[2]}, 
otherwise bi-directional VEinfos will be generated.}

\item{sel.mode.val}{Character. If set NULL, it uses all values in global variables \code{CellTalkDB::kpred.mode}, or
please specify detailed and accurate values in subset of \code{CellTalkDB::kpred.mode}.}

\item{sel.action.effect.val}{Character. If set NULL, it uses all values in global variables \code{CellTalkDB::kpred.action.effect}, or
please specify detailed and accurate values in subset of \code{CellTalkDB::kpred.action.effect}.}
}
\value{
A list.
\itemize{
  \item {\code{cluster.name.A}&\code{cluster.name.B}:} {cluster names involved.}
  \item edges.infos: data.frame that records the edges(the interaction pairs).
  \item vertices.infos: data.frame that records the vertices(the genes).
  \item vertices.apx.type.A: data.frame that records the types(molecular functions) of A in gene pairs formatted as A-B.
  \item vertices.apx.type.B: data.frame that records the types(molecular functions) of B in gene pairs formatted as A-B.
}
}
\description{
This function uses detailed informations about one interaction pair(return value of 
\code{GenerateMapDetailOnepairClusters()}), to generate data for drawing relation plot.
}
\details{
This function only uses actions that are already known, i.e. recorded thoroughly in some databases,
to generate formatted data structure(with vertices and edges).

In vertices, all gene informations are well recorded, and every gene is given one unique ID.

In edges, it uses unique vertices IDs to contruct the linkes, and records mode and action.effect for every link.
}
