% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed-view-in-targets.R
\name{GetResult.PlotOnepairClusters.GeneCrosstalk}
\alias{GetResult.PlotOnepairClusters.GeneCrosstalk}
\title{[TODO]}
\usage{
GetResult.PlotOnepairClusters.GeneCrosstalk(
  VEinfos,
  fgenes.remapped.all,
  direction.A.to.B = TRUE,
  colnames.to.cmp = c("LogFC", "p_val_adj"),
  range.to.use = list(LogFC = c(-Inf, Inf), p_val_adj = c(-Inf, Inf)),
  formula.to.use = list(Tool.formula.onLogFC.default, Tool.formula.onPValAdj.default),
  axis.order.xy = c("AlphaBet", "AlphaBet"),
  axis.order.xy.decreasing = c(TRUE, TRUE),
  nodes.colour.seq = c("#00809D", "#EEEEEE", "#C30000"),
  nodes.colour.value.seq = c(0, 0.5, 1),
  nodes.size.range = c(2, 8)
)
}
\value{
[TODO]
}
\description{
[TODO]
}
\details{
[TODO]
}
