% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-in-targets.R
\name{GetResult.SummaryOnepairClusters}
\alias{GetResult.SummaryOnepairClusters}
\title{Get result of analysis on one pair}
\usage{
GetResult.SummaryOnepairClusters(
  onepair.gmoc,
  show.exprs.change = c("Cup.Dup", "Cup.Ddn", "Cdn.Dup", "Cdn.Ddn"),
  show.action.effects = c("A-->B", "A<--B", "A--|B", "A|--B", "A--oB", "Ao--B",
    "A---B"),
  plot.axis.x.name = NULL,
  plot.axis.y.name = NULL
)
}
\arguments{
\item{onepair.gmoc}{List. Return value of \code{\link{GenerateMapDetailOnepairClusters}}.}

\item{show.exprs.change}{Character. Use exprssion level change of clusters to select part of data to be shown.}

\item{show.action.effects}{Character. Select some action effects to be put in the result.}

\item{plot.axis.x.name}{Character. X-axis name when plot is shown.}

\item{plot.axis.y.name}{Character. Y-axis name when plot is shown.}
}
\value{
A list. Use \code{Tool.ShowGraph()} to see the \bold{plot}, \code{Tool.WriteTables()} to save the result \bold{table} in .csv files.
\itemize{
  \item plot: the object of \pkg{ggplot2}.
  \item table: a list of \code{data.frame}.
}
}
\description{
This function focuses on one interaction pair and its detailed information(expression changes, 
gene-gene action effects), and get result from it.
}
\examples{
# select only Cup.Dup
\dontrun{
GetResult.SummaryOnepairClusters(
  onepair.gmoc,
  show.exprs.change = c("Cup.Dup")
)
}



}
