% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullView.R
\name{GetResultFullView}
\alias{GetResultFullView}
\title{Get Result for Network Analysis}
\usage{
GetResultFullView(
  object,
  show.clusters.in.x = NULL,
  show.clusters.in.y = NULL,
  sel.cluster.group.method = "all",
  power.apply.func = NULL,
  cnt.apply.func = NULL,
  plot.power.range = NULL,
  plot.cnt.range = NULL,
  nodes.size.range = c(1, 6),
  nodes.colour.seq = c("#00809D", "#EEEEEE", "#C30000"),
  nodes.colour.value.seq = c(0, 0.5, 1),
  plot.axis.x.name = "clusters-x",
  plot.axis.y.name = "clusters-y"
)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{show.clusters.in.x}{Clusters chosen to show in x-axis, corresponding to paramter \code{sel.clusters.X} in \code{\link{AnalyzeInterInFullView}}.}

\item{show.clusters.in.y}{Clusters chosen to show in y-axis, corresponding to paramter \code{sel.clusters.Y} in \code{\link{AnalyzeInterInFullView}}.}

\item{sel.cluster.group.method}{Options are 'all', 'diagonal' and 'diagonal-2'. It controls 
'all' is recommended for most cases, and 
'diagonal*' is recommended when same restrictions are applied to clusters in X and Y.}

\item{power.apply.func}{The function used to modify the power value, usually \code{log1p}, etc.}

\item{cnt.apply.func}{The function used to modify the count value, usually \code{sqrt} to avoid extremes problem.}

\item{plot.power.range}{This will extend the plotting range beyond actual value range for power.}

\item{plot.cnt.range}{This will extend the plotting range beyond actual value range for count.}

\item{nodes.size.range}{2 numbers that specifies the range of sizes of the nodes shown in the graph.}

\item{nodes.colour.seq}{Given colours will be used to generate colour gradient for plotting.}

\item{nodes.colour.value.seq}{Numeric. If set NULL, colours given in \code{nodes.colour.seq} will be evenly placed.
Otherwise, numeric values with the same length of \code{nodes.colour.seq} should be given to specify the positions corresponding to each colour.
See parameter \code{values} in \code{ggplot2::scale_colour_gradientn} for details.}

\item{plot.axis.x.name}{X-axis name when plotting.}

\item{plot.axis.y.name}{Y-axis name when plotting.}
}
\value{
List. Use \code{Tool.ShowPlot()} to see the \bold{plot}, \code{Tool.WriteTables()} to save the result \bold{table} in .csv files.
\itemize{
  \item plot: the object of \pkg{ggplot2}.
  \item table: a list of \code{data.frame}.
}
}
\description{
This function summarizes the result of network analysis in graph and table, and gives clues to 
most active intercellular communications.
}
