% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-part1.R
\name{GetResultPieActionEffect}
\alias{GetResultPieActionEffect}
\title{Give composition proportion of action effect}
\usage{
GetResultPieActionEffect(
  object,
  limits.exprs.change = kexprs.change,
  limits.ext.action.effect = kpred.ext.action.effect,
  color.ext.action.effect = kpred.color.ext.effect,
  legend.title = "Action Effect",
  legend.title.size = element_text(size = 14),
  legend.key.size = unit(7, "mm"),
  legend.text.size = element_text(size = 12),
  legend.box.margin = margin(0, 0, 0, 6),
  show.note = TRUE,
  note.label.size = 3,
  note.lineheight = 1,
  note.postion.xy = c(0.5, 0.5),
  note.hjust = 0.5,
  note.vjust = 0.5
)
}
\arguments{
\item{object}{[TODO]}

\item{limits.exprs.change}{Character. Use exprssion level change of clusters to select part of data to be shown.}

\item{limits.ext.action.effect}{[TODO]}

\item{color.ext.action.effect}{[TODO]}

\item{legend.title}{It sets the attributes of legend title, and should be \code{element_text()}.}

\item{legend.key.size}{The size of keys in legend. It should be in unit format, see \code{?unit} for further help.}

\item{legend.text.size}{It sets the attributes of legend annotation texts, and should be \code{element_text()}.}

\item{legend.box.margin}{It sets the margin of legend box, and should be \code{margin()}, see \code{?margin} for further help.}

\item{show.note}{[TODO]}

\item{note.label.size}{[TODO]}

\item{note.lineheight}{[TODO]}

\item{note.postion.xy}{[TODO]}

\item{note.hjust}{[TODO]}

\item{note.vjust}{[TODO]}

\item{color.rev.action.effect}{[TODO]}
}
\value{
A list. Use \code{Tool.ShowGraph()} to see the \bold{plot}, \code{Tool.WriteTables()} to save the result \bold{table} in .csv files.
\itemize{
  \item plot: the object of \pkg{ggplot2}.
  \item table: a list of \code{data.frame}.
}
}
\description{
This function focuses on one interaction pair and its detailed information(expression changes, 
gene-gene action effects), and get result from it.
}
\examples{
# select only Xup.Yup
\dontrun{
GetResult.SummaryOnepairClusters(
  onepair.gmoc,
  limits.exprs.change = c("Xup.Yup")
)
}



}
