% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-part1.R
\name{GetResultPieActionMode}
\alias{GetResultPieActionMode}
\title{Give composition proportion of action mode}
\usage{
GetResultPieActionMode(
  object,
  limits.exprs.change = c("Xup.Yup", "Xup.Ydn", "Xdn.Yup", "Xdn.Ydn"),
  limits.action.mode = NULL,
  color.action.mode = NULL,
  legend.title = "Action Mode",
  legend.title.size = element_text(size = 16),
  legend.key.size = unit(7, "mm"),
  legend.text.size = element_text(size = 12),
  legend.box.margin = margin(0, 0, 0, 6),
  caption.label.size = 12,
  caption.label.x = 0.5,
  caption.label.y = 0.8,
  caption.label.hjust = 0.5,
  caption.label.vjust = 1
)
}
\arguments{
\item{object}{[TODO]}

\item{limits.exprs.change}{Character. Use exprssion level change of clusters to select part of data to be shown.}

\item{limits.action.mode}{Character. Select some action effects to be put in the result.}

\item{legend.title}{It sets the content of legend title.}

\item{legend.title.size}{It sets size of legend title.}

\item{legend.key.size}{The size of keys in legend.}

\item{legend.text.size}{It sets size of legend text.}

\item{legend.box.margin}{It sets the margin of legend box, and should be \code{margin()}, see \code{?margin} for further help.}

\item{caption.label.size}{It sets the size of caption label.}

\item{caption.label.x}{It sets the position of caption label in x-axis.}

\item{caption.label.y}{It sets the position of caption label in y-axis.}

\item{caption.label.hjust}{It sets the alignment type of caption label in horizontal level, see \code{vignette("ggplot2-specs")}
for further help.}

\item{caption.label.vjust}{It sets the alignment type of caption label in vertical level, see \code{vignette("ggplot2-specs")}
for further help.}

\item{use.prioritized.mode}{If TRUE, for one gene pair, it gets only one of belonging multiple mode to be reserved.
Or, use all mode with no priority.}
}
\description{
This function focuses on one interaction pair and its detailed information(expression changes, 
gene-gene action mode), and get result from it.
}
