% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-part2.R
\name{GetResultTgCrosstalk}
\alias{GetResultTgCrosstalk}
\title{Filter Important Gene Pairs in One Interaction}
\usage{
GetResultTgCrosstalk(
  object,
  direction.X.to.Y = NULL,
  colnames.to.cmp = c("LogFC", "PVal"),
  range.to.use = list(LogFC = c(-Inf, Inf), PVal = c(-Inf, Inf)),
  plot.X.to.Y = TRUE,
  axis.order.xy = c("AlphaBet", "AlphaBet"),
  axis.order.xy.decreasing = c(TRUE, TRUE),
  plot.font.size.base = 12,
  nodes.colour.seq = c("#00809D", "#EEEEEE", "#C30000"),
  nodes.colour.value.seq = c(0, 0.5, 1),
  nodes.size.range = c(2, 8),
  axis.text.x.pattern = element_text(angle = 30, hjust = 1)
)
}
\arguments{
\item{object}{A \code{InterCell} object, which is created by \code{\link{CreateInterCellObject}}.}

\item{direction.X.to.Y}{Options are 'NULL', 'TRUE', 'FALSE'. It selects subset of data based on direction.
The 'NULL' will keep 2-way interacting pairs, 'TRUE' keeps the X-to-Y pairs and 'FALSE' keeps the Y-to-X pairs. 
See details for help.}

\item{colnames.to.cmp}{The colnames to be used as evaluation params, currently only 2 options ('LogFC', 'PVal') are supported.
The 1st one will be plotted by different size of nodes, and 2nd one will be distinguished by colour of nodes.}

\item{range.to.use}{It specifies the user specified ranges for evaluation params.}

\item{plot.X.to.Y}{The clusters drawn in x-axis and y-axis are in default aligned with the network analysis.
If set FALSE, switch the clusters drawn in x-axis and y-axis.}

\item{axis.order.xy}{It determines how the gene names will be ordered in the axis when plotting.}

\item{axis.order.xy.decreasing}{It determines whether the orders are of decreasing pattern or increasing pattern.}

\item{plot.font.size.base}{It gives the font size of texts such as labels and titles.}

\item{nodes.colour.seq}{It specifies the colour sequence of the nodes.}

\item{nodes.colour.value.seq}{It is along with the param \code{nodes.colour.seq}, and changes the colour expansion.}

\item{nodes.size.range}{It specifies the size range of the nodes.}

\item{axis.text.x.pattern}{It defines the axis text style in x-axis.}
}
\value{
List. Use \code{Tool.ShowPlot()} to see the \bold{plot}, \code{Tool.WriteTables()} to save the result \bold{table} in .csv files.
\itemize{
  \item plot: the object of \pkg{ggplot2}.
  \item table: a list of \code{data.frame}.
}
}
\description{
This function help filter and rank the important gene pairs by evaluating both power and confidence, 
which is calculated from 'LogFC' and 'PVal' respectively.
}
\details{
The meaning for 2 used values is:
\itemize{
  \item LogFC: the log fold change, which indicates the relative gene expression.
  \item PVal: the confidence of discovering the gene as differently expressed genes. 
              If it is generated from Seurat, it is orginally calculated by bonferroni correction.
}

Illustration for \code{direction.X.to.Y}:
When running this function, gene pairs have been combined with their actions. As a result, gene pairs 
get to have direction for their action, e.g. IL6->IL6R, which means IL6 gets to activate IL6R, and the direction
should be IL6 to IL6R. Suppose IL6 is expressed by cell cluster X, IL6R is expressed by cell cluster Y, then 
IL6->IL6R will be reserved if \code{direction.X.to.Y} is set either 'NULL' or 'TRUE', but not 'FALSE'. 
The cluster X and Y is aligned with what users specified in \code{link{AnalyzeInterInFullView}}, and X corresponds to 
those clusters shown in x-axis and Y corresponds to those in y-axis. More closely, the X and Y are corresponding to 
\code{cluster.x} and \code{cluster.y} given by \code{\link{FetchInterOI}}.
}
