% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-part3.R
\name{GetResultTgSpecificity}
\alias{GetResultTgSpecificity}
\title{Summary special genes in cluster groups}
\usage{
GetResultTgSpecificity(
  object,
  show.uq.cnt.range = integer(),
  show.uq.cnt.merged = TRUE,
  select.genepairs = character(),
  select.genepairs.method = "random",
  select.by.method.pairs.limit = 10,
  select.by.method.decreasing = TRUE,
  show.topn.inside.onepair = +Inf,
  prioritize.cluster.groups = character(),
  display.conv.or.rev = TRUE,
  grid.plot.ncol = 1,
  barplot.or.dotplot = FALSE,
  plot.font.size.base = 12,
  axis.text.x.pattern = element_text(angle = 90, vjust = 0.5, hjust = 1),
  bar.facet.scales = "free_x",
  bar.facet.space = "free_x",
  bar.facet.text.x = element_text(size = 8, colour = "black"),
  bar.facet.background = element_rect(fill = "lightgrey", colour = "white"),
  bar.colour = character(),
  bar.width = 0.8,
  dot.range.to.use = list(LogFC = c(-Inf, +Inf), PVal = c(-Inf, +Inf)),
  dot.colour.palette = scale_colour_gradientn(name = "PVal", colours = c("#00809D",
    "#EEEEEE", "#C30000"), values = c(0, 0.5, 1)),
  dot.size.range = c(2, 8),
  dot.y.order.in.alphabet = TRUE,
  ...
)
}
\arguments{
\item{object}{[TODO]}

\item{show.uq.cnt.range}{Integer. It defines the range of \code{uq.cnt} that is going to be used in downstream analysis, and 
the default setting is to use all valid \code{uq.cnt}.}

\item{show.uq.cnt.merged}{Logic. If set TRUE, then gene pairs of different \code{uq.cnt} will be merged in analysis, or gene pairs 
will be grouped by their \code{uq.cnt} and be plotted accordingly.}

\item{select.genepairs}{Character. It specificly gives the gene pairs that are going to be analyzed.}

\item{select.genepairs.method}{Character. It has options: "random", "logfc-sum". It works only when no specific gene pairs are given in 
parameter \code{select.genepairs}. Method "random" will randomly pick up some gene pairs. Method "logfc-sum" will calculate the sum of LogFCs of 
the 2 genes in every gene pairs, and order them by the calculated values.}

\item{select.by.method.pairs.limit}{Integer. It defines the maximum number of gene pairs that are selected by any method.}

\item{select.by.method.decreasing}{Logic. It works for method "logfc-sum". If TRUE, result will be ordered in decreasing way, otherwise the increasing direction.}

\item{show.topn.inside.onepair}{Integer. One gene pairs get to be shared by several cluster groups. By setting this parameter, 
only the top ranked some gene pairs will be finally shown in result. The default value is +Inf, which preserves all result.}

\item{prioritize.cluster.groups}{Character. It defines the most concerning cluster groups, and those cluster groups given in this parameter, will be 
finally plotted from the most left-side to right, and as such, it is called prioritizing.}

\item{display.conv.or.rev}{[TODO] Is the display of cluster interaction the same arrange as VEinfos(direction.A.to.B)
conv, the same, rev, the reverse format. For example, cluster.A -> cluster.B in VEinfo, so, if conv, display like cluster.A~cluster.B, or as cluster.B~cluster.A
gene pairs will be aligned as cluster order changed.}

\item{grid.plot.ncol}{[TODO]show.uq.cnt.merged = FALSE, then it will be used}

\item{barplot.or.dotplot}{[TODO]}

\item{plot.font.size.base}{Numeric. It defines the font size of texts such as labels and titles.}

\item{axis.text.x.pattern}{It defines the axis text style in x-axis.}

\item{bar.facet.scales}{It controls the scales that facet uses, and gets 4 options as defined by \pkg{ggplot2}: "fixed", "free", "free_x", "free_y".}

\item{bar.facet.space}{It controls the space allocating strategy that facet uses, and gets 4 options as defined by \pkg{ggplot2}: "fixed", "free", "free_x", "free_y".}

\item{bar.facet.text.x}{It defines the facet labeling text on the top horizontal position.}

\item{bar.facet.background}{It defines the background style of labellers in facet way.}

\item{bar.colour}{Character. It gives all optional colours that plotting bars get to use. If no specific colour is given, then the 
built-in 20 kinds of colours will be automatically used.}

\item{bar.width}{Numeric. It defines the bar width.}

\item{dot.range.to.use}{[TODO]}

\item{dot.colour.palette}{[TODO]}

\item{dot.size.range}{[TODO]}

\item{dot.y.order.in.alphabet}{[TODO]}

\item{...}{Other parameter that can be passed to select by method functions.}
}
\value{
List. Use \code{Tool.ShowPlot()} to see the \bold{plot}, \code{Tool.WriteTables()} to save the result \bold{table} in .csv files.
\itemize{
  \item plot: the object of \pkg{ggplot2}.
  \item table: a list of \code{data.frame}.
}
}
\description{
This function is to summary special genes and their specific expressing attributes.
}
