% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TgView-prep.R
\name{SelectInterSubset}
\alias{SelectInterSubset}
\title{Select part of gene pairs}
\usage{
SelectInterSubset(
  object,
  direction.X.to.Y = NULL,
  sel.exprs.change = c("Xup.Yup", "Xup.Ydn", "Xdn.Yup", "Xdn.Ydn"),
  sel.some.genes.X = NULL,
  sel.some.genes.Y = NULL,
  sel.genes.option = "intersect",
  sel.gene.pairs = NULL,
  sel.some.gene.pairs.colnames = c("inter.GeneName.A", "inter.GeneName.B"),
  sel.mode.val = NULL,
  sel.action.effect.val = NULL,
  sel.mode.action.option = "intersect"
)
}
\arguments{
\item{object}{[TODO]}

\item{direction.X.to.Y}{[TODO]}

\item{sel.exprs.change}{Character. It selects the expression change status that gene pairs can be. It has total 4 options:
"Xup.Yup", "Xup.Ydn", "Xdn.Yup", "Xdn.Ydn", which are defined in global variables \code{kexprs.change}.}

\item{sel.some.genes.X}{Character. It selects some genes in cluster A(A is defined in parameter \code{VEinfos} and use \code{VEinfos$cluster.name.A} to fetch that).}

\item{sel.some.genes.Y}{Character. It selects some genes in cluster B(A is defined in parameter \code{VEinfos} and use \code{VEinfos$cluster.name.B} to fetch that).}

\item{sel.genes.option}{Character. Its allowed values are "intersect" and "union". It defines the way that merges the result of subset selected by some given genes.
The default way is to intersect the subsets, and the other option is union.}

\item{sel.gene.pairs}{Data.frame. It is at-least-2-column data.frame, which records gene pairs with each column settling 
one of the participated genes. The 2 required column names need to be specified by another parameter \code{sel.some.gene.pairs.colnames}.}

\item{sel.some.gene.pairs.colnames}{Character of length 2. It strictly specifies the column names that records the genes of given gene pairs, 
and it also implies the direction goes from the first column to the second (AtoB). So, make sure putting genes in their proper positions.}

\item{sel.mode.val}{Character. If set NULL, it uses all values in global variables \code{InterCellDB::kpred.action.mode}, or
please specify detailed and accurate values in subset of \code{InterCellDB::kpred.action.mode}.}

\item{sel.action.effect.val}{Character. If set NULL, it uses all values in global variables \code{InterCellDB::kpred.action.effect}, or
please specify detailed and accurate values in subset of \code{InterCellDB::kpred.action.effect}.}

\item{sel.mode.action.option}{Character. Its allowed values are "intersect" and "union". It defines the way that merges the result of subset selected by mode 
and subset selected by action.effect. The default way is to intersect the subsets, and the other option is union.}
}
\description{
This function is to further select subset of interactions in given 2-cell group.
}
\details{
The whole list of mode or action.effect is defined as global variable that is given within the package.
kpred.action.mode defines 9 modes, while kpred.action.effect defines 4 action effects.
As the action database given now is not completed well, it is recommended to select upon
action.effect, while leaving all different modes preserved.
}
