% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed-view-in-targets.R
\name{TrimVEinfos}
\alias{TrimVEinfos}
\title{[TODO]}
\usage{
TrimVEinfos(VEinfos, sel.mode.val = NULL, sel.action.effect.val = NULL)
}
\arguments{
\item{VEinfos}{List. It contains informations about vertices and edges, and is exactly return value of
\code{GenerateVEinfos()}.}

\item{sel.mode.val}{Character. If set NULL, it uses all values in global variables \code{CellTalkDB::kpred.mode}, or
please specify detailed and accurate values in subset of \code{CellTalkDB::kpred.mode}.}

\item{sel.action.effect.val}{Character. If set NULL, it uses all values in global variables \code{CellTalkDB::kpred.action.effect}, or
please specify detailed and accurate values in subset of \code{CellTalkDB::kpred.action.effect}.}
}
\description{
[TODO]
}
\details{
[TODO]
}
