% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed-view-in-targets.R
\name{TrimVEinfos}
\alias{TrimVEinfos}
\title{Trim the vertices and edges infos}
\usage{
TrimVEinfos(VEinfos, sel.mode.val = NULL, sel.action.effect.val = NULL)
}
\arguments{
\item{VEinfos}{List. It contains informations about vertices and edges, and is exactly return value of
\code{GenerateVEinfos()}.}

\item{sel.mode.val}{Character. If set NULL, it uses all values in global variables \code{CellTalkDB::kpred.mode}, or
please specify detailed and accurate values in subset of \code{CellTalkDB::kpred.mode}.}

\item{sel.action.effect.val}{Character. If set NULL, it uses all values in global variables \code{CellTalkDB::kpred.action.effect}, or
please specify detailed and accurate values in subset of \code{CellTalkDB::kpred.action.effect}.}
}
\description{
This function uses mode and action.effect as the restrictions when users select part of the result.
}
\details{
The whole list of mode or action.effect is defined as global variable that is given within the package.
kpred.mode defines 9 modes, while kpred.action.effect defines 4 action effects.
As the action database given now is not completed well, it is recommended to select upon
action.effect, while leaving all different modes preserved.
}
